### 2019-05-04 #1431 @whedon
[PRE REVIEW]: emg3d: A multigrid solver for 3D electromagnetic diffusion.
**Submitting author:** @prisae (<a href="http://orcid.org/0000-0002-8575-2484">Dieter Werthmüller</a>)
**Repository:** <a href="https://github.com/empymod/emg3d" target ="_blank">https://github.com/empymod/emg3d</a>
**Version:** v0.5.0
