### 2019-05-05 #1431 @prisae
Re reviewer: Anyone into geophysics and python; electromagnetics would be a plus, I guess.

Some people from your list which fit the description: @rowanc1, @leouieda, @WalBytes , @vidalgp, @kbevers, @jkmacc-LANL, @whimian, @rbeucher, @d-chambers.
