### 2019-05-17 2cf30cb
Change Travis dist
