+++ .travis.yml
@@ -4,11 +4,9 @@ language: python
 
 matrix:
   include:
-    - name: "Channel `defaults`; jitted"
-      python: 3.7
+    - python: 3.7  # => 3.7 with defaults; mkl
       env: PYTHON=3.7 PCKGS="mkl mkl-service numba" CHAN="defaults" INST=""
-    - name: "Channel `conda-forge`; not jitted"
-      python: 3.7
+    - python: 3.7  # => 3.7 with conda-forge; IPython, discretize, no JIT
       env: PYTHON=3.7 PCKGS="numba IPython" NUMBA_DISABLE_JIT=1 CHAN="conda-forge" INST="discretize"
 
 install:
