### 2019-05-17 e0689c1
Add some notes
