+++ .travis.yml
@@ -2,6 +2,9 @@ dist: xenial  # Required for Python >= 3.7
 
 language: python
 
+# Note: the -python flag in the matrix has no influence, as we use miniconda to
+# install Python, which uses the env-variable PYTHON. Having it here is just
+# bookkeeping, so it prints the same version in the Travis info.
 matrix:
   include:
     - python: 3.7  # => 3.7 with defaults; mkl
