### 2019-05-17 5521ba3
Change dist on Travis.
Change dist on Travis. This way we can use the Python version 3.7, which will display the correct version on the Travis page. However, this is purely cosmetic, as we don't actually use the Travis Python version, but miniconda.
