+++ .travis.yml
@@ -1,11 +1,15 @@
-# http://conda.pydata.org/docs/travis.html
+dist: xenial  # Required for Python >= 3.7
+
 language: python
 
+# Note: the -python flag in the matrix has no influence, as we use miniconda to
+# install Python, which uses the env-variable PYTHON. Having it here is just
+# bookkeeping, so it prints the same version in the Travis info.
 matrix:
   include:
-    - python: 3.6  # => 3.7
+    - python: 3.7  # => 3.7 with defaults; mkl
       env: PYTHON=3.7 PCKGS="mkl mkl-service numba" CHAN="defaults" INST=""
-    - python: 3.6  # => 3.7 with conda-forge; IPython, discretize, no JIT
+    - python: 3.7  # => 3.7 with conda-forge; IPython, discretize, no JIT
       env: PYTHON=3.7 PCKGS="numba IPython" NUMBA_DISABLE_JIT=1 CHAN="conda-forge" INST="discretize"
 
 install:
