+++ docs/mgwhat.rst
@@ -1,5 +1,3 @@
-.. _MGWhat:
-
 Multi-what?
 ###########
 
@@ -26,7 +24,8 @@ the full matrix. This is how it achieves its relatively low memory consumption.
 To solve the system, it solves for all fields adjacent to one node, moves then
 to the next node, and so on until it reaches the last node, see :numref:`Figure
 %s <smoother>`, where the red lines indicate the fields which are solved
-simultaneously per step.
+simultaneously per step (the fields on the boundaries are never calculated, as
+they are assumed to be 0).
 
 .. figure:: _static/schematics2.svg
    :scale: 80 %
