### 2019-05-20 d653640
Add references: Mulder 2011, Fedorenko 1964
