### 2019-05-20 #1463 @jedbrown
@akelbert @emersodb @lukeolson :wave:  Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [emg3d repository](https://github.com/empymod/emg3d)). I'll be watching this thread if you have any questions.
