### 2019-05-20 #1463 @prisae
Great to see the review started! I repeat a question here that I asked in the original issue (https://github.com/openjournals/joss-reviews/issues/1431#issuecomment-489406698): "I saw one typo and some of the reference-bracketing in the compiled version did not turn out as I intended. Is there a way to edit them now or shall I wait for the reviews first?"
