### 2019-05-20 #1463 @danielskatz
You can make changes in the paper.md and paper.bib - after you have done so, enter a new comment here containing ```@whedon generate pdf``` to rebuild the pdf
