### 2019-05-21 2027a1a
Numba for fx prolong working
