### 2019-05-22 a11e0ae
Numba for fy/fz prolong working
