### 2019-05-22 ee5bc1e
Combine prolong_fx/y/z
