+++ emg3d/njitted.py
@@ -2026,72 +2026,53 @@ def restrict_weights(vectorN, vectorCC, h, cvectorN, cvectorCC, ch):
 
 # Prolongation
 @nb.njit(**_numba_setting)
-def prolong(cefield, indices, norm_distances):
-    """Interpolating coarse grid e-field to fine grid.
+def prolong_field(vectorN1, vectorN2, cefield, points, nC):
+    """Bilinear interpolation in 2D planes.
 
-    This is a custom version of
-    :class:`scipy.interpolate.RegularGridInterpolator`; see those docs for more
-    information.
-
-    The customization makes the following assumption and adjustments:
-
-    - Only supports 2D;
-    - ``bounds_error=False``;
-    - ``fill_value=None``;
-    - changes in order to jitt it;
-    - re-arrangement so only the necessary things have to be done inside the
-      loop in :func:`emg3d.solver.prolongation`;
-    - No sanity checks are carried out; the input must be as expected.
+    Prolongate the e-field normal to the plane by looping over each slice,
+    interpolating within the plane.
 
 
     Parameters
     ----------
-    cefield : Field
-        Coarse grid electric field; ``emg3d.utils.Field`` instance.
+    vectorN1, vectorN2 : ndarray
+        Cell edges of the coarse grid in the two plane-directions.
 
-    indices, norm_distances : lists of two 1D ndarrays
-        Output of :func:`prolong_init`.
+    cefield : ndarray
+        Coarse grid electric field directed normal to plane.
+
+    points : ndarray
+        2D array containing the plane-coordinates to interpolate.
+
+    nC : int
+        Number of coarse grid cells in the direction of the e-field. Must be
+        one of cefield.shape, otherwise a zero field is returned.
+
 
     Returns
     -------
-    field : ndarray
-        Coarse grid electric field interpolated at fine grid locations.
-
-    Notes
-    -----
-    We set ``bounds_error=False`` and ``fill_value=None`` in
-    :class:`scipy.interpolate.RegularGridInterpolator` to extrapolate fine grid
-    points which are outside the coarse grid. The fine grid points are,
-    theoretically, never outside the coarse grid points. However, the
-    restriction mesh is created with the distances between points. This can
-    lead to fine grid points slightly outside coarse grid points (in the orders
-    of 1e-10 m) with very big distances (grids of roughly over 1e6 m). So
-    basically nothing, but it would still cause an error in the interpolation.
+    efield : ndarray
+        Fine grid electric field normal to plane.
 
     """
-    # Number of new points.
-    ni = indices[:, 0].size
+    # Pre-allocate interpolated field.
+    efield = np.zeros((nC, points.shape[0]), dtype=np.complex128)
 
-    # Find relevant fine grid field; each i and i+1 represents an edge.
-    field = np.zeros(ni, dtype=np.complex128)
-    for i in range(4):
-        weight = np.ones(ni)
-        edge_indices = (indices[:, 0] + i // 2, indices[:, 1] + i % 2)
+    # Initialize interpolation.
+    rgi_inp = prolong_init((vectorN1, vectorN2), points)
 
-        # Calculate the weights.
-        for ii in range(2):
-            ei = edge_indices[ii]
-            ind = indices[:, ii]
-            yi = norm_distances[:, ii]
+    # Bilinear interpolation for each plane/slice.
+    if nC == cefield.shape[0]:   # x-directed.
+        for ic in range(nC):
[…]
