### 2019-05-22 54fef35
Replace RegularGridInterpolator
Replace `scipy.interpolate.RegularGridInterpolator` with a custom version of it, jitted.

This can speed-up the prolongation massively for small grids (up to 10 times faster) and moderately for huge grids (1.15 times faster), which in turns speeds-up the entire multigrid solver. However, this is only visible if MG is used as a solver, not as a pre-conditioner.

The customization makes the following assumption and adjustments:
- Only supports 2D;
- `bounds_error=False`;
- `fill_value=None`;
- changes in order to jitt it;
- re-arrangement so only the necessary thi
[…]
