+++ emg3d/njitted.py
@@ -2022,3 +2022,162 @@ def restrict_weights(vectorN, vectorCC, h, cvectorN, cvectorCC, ch):
         wr[i] *= cvectorCC[i]-vectorCC[2*i]
 
     return wl, w0, wr
+
+
+# Prolongation
+@nb.njit(**_numba_setting)
+def prolong_field(vectorN1, vectorN2, cefield, points, nC):
+    """Bilinear interpolation in 2D planes.
+
+    Prolongate the e-field normal to the plane by looping over each slice,
+    interpolating within the plane.
+
+
+    Parameters
+    ----------
+    vectorN1, vectorN2 : ndarray
+        Cell edges of the coarse grid in the two plane-directions.
+
+    cefield : ndarray
+        Coarse grid electric field directed normal to plane.
+
+    points : ndarray
+        2D array containing the plane-coordinates to interpolate.
+
+    nC : int
+        Number of coarse grid cells in the direction of the e-field. Must be
+        one of cefield.shape, otherwise a zero field is returned.
+
+
+    Returns
+    -------
+    efield : ndarray
+        Fine grid electric field normal to plane.
+
+    """
+    # Pre-allocate interpolated field.
+    efield = np.zeros((nC, points.shape[0]), dtype=np.complex128)
+
+    # Initialize interpolation.
+    rgi_inp = prolong_init((vectorN1, vectorN2), points)
+
+    # Bilinear interpolation for each plane/slice.
+    if nC == cefield.shape[0]:   # x-directed.
+        for ic in range(nC):
+            prolong(efield[ic, :], cefield[ic, :, :], *rgi_inp)
+    elif nC == cefield.shape[1]:  # y-directed.
+        for ic in range(nC):
+            prolong(efield[ic, :], cefield[:, ic, :], *rgi_inp)
+    elif nC == cefield.shape[2]:  # z-directed.
+        for ic in range(nC):
+            prolong(efield[ic, :], cefield[:, :, ic], *rgi_inp)
+
+    return efield
+
+
+@nb.njit(**_numba_setting)
+def prolong_init(xy, xi):
+    """Initialization for `prolong`.
+
+    See :func:`prolong` for details.
+
+    Parameters
+    ----------
+    xy : tuple of two 1D arrays (x, y)
+        x and y are the grid points in x and y.
+
+    xi : ndarray of shape (..., 2)
+        The coordinates to sample the gridded data at.
+
+    Returns
+    -------
+    indices : list of two 1D ndarrays
+        Indices for interpolation from xy to xi.
+
+    norm_distances : list of two 1D ndarrays
+        Normalized distances for interpolation from xy to xi.
+
+    """
+
+    # Find relevant edges between which xi are situated.
+    indices = np.zeros(xi.shape, dtype=np.int16)
+
+    # Compute distance to lower edge in unity units.
+    norm_distances = np.zeros(xi.shape)
+
+    # Iterate through x, y.
+    for i in range(2):
+
+        ind = np.searchsorted(xy[i], xi[:, i]) - 1
+        ind[ind < 0] = 0
+        ind[ind > xy[i].size - 2] = xy[i].size - 2
+
+        indices[:, i] = ind
+
+        normd = ((xi[:, i] - xy[i][ind]) / (xy[i][ind + 1] - xy[i][ind]))
+        norm_distances[:, i] = normd
+
+    return indices, norm_distances
+
+
+@nb.njit(**_numba_setting)
+def prolong(efield, cefield, indices, norm_distances):
+    """Interpolating coarse grid e-field to fine grid.
+
+    This is a custom version of
+    :class:`scipy.interpolate.RegularGridInterpolator`; see those docs for more
+    information.
+
[…]
