### 2019-05-23 7993df1
Revert previous commit
My metrics of measuring the speed-up were wrong. While there was a
speed-up for small grids, there was actually a slow-down for big grids.
So not working.
