### 2019-05-23 697760f
Factor krylov out of solver jk-I.
First re-factoring in a series to improve the code and make it easier to
test and maintain.

- Factor `krylov` out of `solver`.
- Introduce `var.cprint`.
