+++ emg3d/solver.py
@@ -34,7 +34,7 @@ from . import utils
 from . import njitted
 
 __all__ = ['solver', 'multigrid', 'smoothing', 'restriction', 'prolongation',
-           'residual', 'MGParameters']
+           'residual', 'krylov', 'MGParameters']
 
 
 def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
@@ -300,154 +300,57 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
     )
 
     # Print all parameters if verbose.
-    if var.verb > 1:
-        print(f"\n:: emg3d START :: {var.time.now} ::\n")
-        print(var)
+    var.cprint(f"\n:: emg3d START :: {var.time.now} ::\n", 1)
+    var.cprint(var, 1)
 
     # Get efield
-    do_run = True     # Flag whether or not to run the solver.
-    do_return = True  # Flag whether or not to return the field.
     if efield is None:
         # If not provided, initiate an empty one.
         efield = utils.Field(grid)
+
+        # Set flag to return the field.
+        do_return = True
     else:
         # If provided, take the conjugate (see return statement).
         efield.field = efield.field.conjugate()
 
-        # Don't return the field, just inplace.
+        # Set flag to NOT return the field.
         do_return = False
 
         # If efield is provided, check if it is already sufficiently good.
         var.l2 = np.linalg.norm(residual(grid, model, sfield, efield))
         if var.l2 < var.tol*np.linalg.norm(sfield):
-            do_run = False
+            info = f"   > Provided efield already good enough!"
+            var.cprint(info+f":: emg3d END ::\n", 1)
+            return
 
-    # Switch if multigrid or ssl-solver is main solver.
-    if do_run and var.sslsolver:  # `ssl-solver` is main solver.
-
-        # Print header of iteration log.
-        if var.verb > 2:
-            prnt = f"   [hh:mm:ss]     {'error':<15}{'solver':<20} "
-            if var.cycle:
-                prnt += f"{'MG':<11} l s"
-            print(prnt+f"\n")
-
-        # Define matrix operation A x as LinearOperator.
-        def amatvec(efield):
-            """Compute A x for solver; residual is b-Ax = src-amatvec."""
-
-            # Cast current efield to Field instance.
-            efield = utils.Field(grid, efield)
-
-            # Calculate A x.
-            rfield = utils.Field(grid)
-            njitted.amat_x(
-                    rfield.fx, rfield.fy, rfield.fz,
-                    efield.fx, efield.fy, efield.fz, model.eta_x, model.eta_y,
-                    model.eta_z, model.v_mu_r, grid.hx, grid.hy, grid.hz)
-
-            # Return Field instance.
-            return rfield
-
-        # Initiate LinearOperator A x.
-        A = ssl.LinearOperator(
-                shape=(grid.nE, grid.nE), dtype=complex, matvec=amatvec)
-
-        # Define MG pre-conditioner as LinearOperator, if `var.cycle`.
-        def mg_matvec(sfield):
-            """Use multigrid as pre-conditioner."""
-
-            # Cast current fields to Field instances.
-            sfield = utils.Field(grid, sfield)
-            efield = utils.Field(grid)
-
-            # Solve for these fields.
-            multigrid(grid, model, sfield, efield, var)
-
-            # Set first_cycle off to reduce verbosity after first cycle.
-            var._first_cycle = False
-
-            return efield
-
-        # Initiate LinearOperator M.
-        M = None
+    # Print header for iteration log.
+    header = f"   [hh:mm:ss]     {'error':<15}"
[…]
