### 2019-05-23 148c44d
Add test for providing e-field
