### 2019-05-23 da58946
Factor out info and terminate of multigrid jk-II.
- print_cycle_info
- terminate
