+++ emg3d/solver.py
@@ -34,9 +34,11 @@ from . import utils
 from . import njitted
 
 __all__ = ['solver', 'multigrid', 'smoothing', 'restriction', 'prolongation',
-           'residual', 'krylov', 'MGParameters']
+           'residual', 'krylov', 'MGParameters', 'print_cycle_info',
+           'terminate']
 
 
+# MAIN USER-FACING FUNCTION
 def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
            semicoarsening=False, linerelaxation=False, verb=2, **kwargs):
     r"""Solver for 3D CSEM data with tri-axial electrical anisotropy.
@@ -361,6 +363,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         return efield
 
 
+# SOLVERS
 def multigrid(grid, model, sfield, efield, var, **kwargs):
     """Multigrid solver for 3D controlled-source electromagnetic (CSEM) data.
 
@@ -551,55 +554,8 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
 
         else:          # Original grid reached, check termination criteria.
 
-            # Print cycle info if verbose.
-            if var.verb > 2:
-                if var.verb > 3:
-                    print()
-
-                # Print multigrid-cycle visual QC on first cycle.
-                if var._first_cycle:
-
-                    # Cast levels into array, get maximum.
-                    _lvl_all = np.array(var._level_all, dtype=int)
-                    lvl_max = np.max(_lvl_all)
-
-                    # Get levels, multiply by difference to get +/-.
-                    lvl = (_lvl_all[1:] + _lvl_all[:-1])//2+1
-                    lvl *= _lvl_all[1:] - _lvl_all[:-1]
-
-                    # Create info string.
-                    out = [f"       h_\n"]
-                    slen = min(len(lvl), 70)
-                    for cl in range(lvl_max):
-                        out += f"   {2**(cl+1):4}h_ "
-                        out += [" " if abs(lvl[v]) != cl+1 else "\\" if
-                                lvl[v] > 0 else "/" for v in range(slen)]
-                        if cl < lvl_max-1:
-                            out.append("\n")
-
-                    # Print the cycle.
-                    print("".join(out), "\n")
-                    if len(lvl) > 70:
-                        print("  (Cycle-QC restricted to first 70 steps of "
-                              f"{len(lvl)} steps.)\n")
-
-                    # Reset _level_all
-                    var._level_all = [0, ]
-
-                # Print iteration log.
-                if var.sslsolver:  # For multigrid as preconditioner.
-                    print(f"   [{var.time.now}] {l2_last:.3e} "
-                          f"after {20*' '} {var.it:2} {var.cycle}-cycles; "
-                          f"  {var.ldir} {var.rdir}")
-
-                else:              # For multigrid as solver.
-                    print(f"   [{var.time.now}] {l2_last:.3e} "
-                          f"after {var.it:2} {var.cycle}-cycles; "
-                          f"[{l2_last/l2_init:.3e}, {l2_last/l2_prev:.3e}]"
-                          f" {var.ldir} {var.rdir}")
-
-                if var.verb > 3:
-                    print()
+            # Print end-of-cycle info.
+            print_cycle_info(var, l2_last, l2_prev, l2_init)
 
             # Adjust semicoarsening and line relaxation if they cycle.
             if var.rcycle:
@@ -607,39 +563,9 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
[…]
