### 2019-05-23 13b4888
Factor out current lr_dir and sc_dir; jk-III.
- current_sc_dir
- current_lr_dir
- Rename rdir to sc_dir; ldir to lr_dir
