### 2019-05-23 501624a
Add new functions to __all__
