### 2019-05-23 4140956
Replace RegularGridInterpolator.
Replace `scipy.interpolate.RegularGridInterpolator` with a custom
version of it.

The main difference (besides the pre-sets) is that this version allows
to initiate an instance with the coarse and fine grids. This initialize
will calculate the required weights, and it has therefore only to be
done once.

After this, interpolating values from the coarse to the fine grid can be
carried out much faster.

Simplifications in comparison to
`scipy.interpolate.RegularGridInterpolator`

- No sanity checks what-so-ever.
- Only 2D data;
- ``method='linear'``;
- ``bounds_e
[…]
