+++ emg3d/solver.py
@@ -24,9 +24,8 @@ are in the :mod:`emg3d.njitted` as numba-jitted functions.
 # the License.
 
 
+import itertools
 import numpy as np
-from itertools import cycle
-import scipy.interpolate as si
 import scipy.sparse.linalg as ssl
 from dataclasses import dataclass
 
@@ -35,7 +34,8 @@ from . import njitted
 
 __all__ = ['solver', 'multigrid', 'smoothing', 'restriction', 'prolongation',
            'residual', 'krylov', 'MGParameters', 'current_sc_dir',
-           'current_lr_dir', 'print_cycle_info', 'terminate']
+           'current_lr_dir', 'print_cycle_info', 'terminate',
+           'RegularGridProlongator']
 
 
 # MAIN USER-FACING FUNCTION
@@ -918,33 +918,19 @@ def prolongation(grid, efield, cgrid, cefield, sc_dir):
     sc_dir : int
         Direction of semicoarsening (0, 1, 2, or 3).
 
-    Notes
-    -----
-    We set ``bounds_error=False`` and ``fill_value=None`` in
-    :class:`scipy.interpolate.RegularGridInterpolator` to extrapolate fine grid
-    points which are outside the coarse grid. The fine grid points are,
-    theoretically, never outside the coarse grid points. However, the
-    restriction mesh is created with the distances between points. This can
-    lead to fine grid points slightly outside coarse grid points (in the orders
-    of 1e-10 m) with very big distances (grids of roughly over 1e6 m). So
-    basically nothing, but it would still cause an error in the interpolation.
-
     """
 
     # Calculate required points of finer grid.
-    x_pts = grid.gridEx[::grid.nCx, 1:]
-    y_pts = grid.gridEy[:, ::2].reshape(grid.nNz, -1, 2)
-    y_pts = y_pts[:, :grid.nNx, :].reshape(-1, 2)
-    z_pts = grid.gridEz[:grid.nNx*grid.nNy, :2]
+    yz_points = grid.gridEx[::grid.nCx, 1:]
+    xz_points = grid.gridEy[:, ::2].reshape(grid.nNz, -1, 2)
+    xz_points = xz_points[:, :grid.nNx, :].reshape(-1, 2)
+    xy_points = grid.gridEz[:grid.nNx*grid.nNy, :2]
 
     # Interpolate ex in y-z-slices.
+    fn = RegularGridProlongator(cgrid.vectorNy, cgrid.vectorNz, yz_points)
     for ixc in range(cgrid.nCx):
-
         # Bilinear interpolation in the y-z plane
-        fn = si.RegularGridInterpolator(
-                (cgrid.vectorNy, cgrid.vectorNz), cefield.fx[ixc, :, :],
-                bounds_error=False, fill_value=None)
-        hh = fn(x_pts).reshape(grid.vnEx[1:], order='F')
+        hh = fn(cefield.fx[ixc, :, :]).reshape(grid.vnEx[1:], order='F')
 
         # Piecewise constant interpolation in x-direction
         if sc_dir not in [1, 5, 6]:
@@ -954,13 +940,11 @@ def prolongation(grid, efield, cgrid, cefield, sc_dir):
             efield.fx[ixc, :, :] += hh
 
     # Interpolate ey in x-z-slices.
+    fn = RegularGridProlongator(cgrid.vectorNx, cgrid.vectorNz, xz_points)
     for iyc in range(cgrid.nCy):
 
         # Bilinear interpolation in the x-z plane
-        fn = si.RegularGridInterpolator(
-                (cgrid.vectorNx, cgrid.vectorNz), cefield.fy[:, iyc, :],
[…]
