### 2019-05-24 869d223
Further simplification RegGridProlongator
