### 2019-05-24 eef25f7
Refactoring mainly `multigrid`; jk-IV.
- `residual` returns new also `l2_norm`;
- `smoothing` returns new `residual`;
- replaced many print with var.cprint;
- moved `print_gs_info` to `MGParameters`.
