+++ emg3d/solver.py
@@ -317,7 +317,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         efield.field = efield.field.conjugate()
 
         # If efield is provided, check if it is already sufficiently good.
-        var.l2 = np.linalg.norm(residual(grid, model, sfield, efield))
+        _, var.l2 = residual(grid, model, sfield, efield)
         if var.l2 < var.tol*np.linalg.norm(sfield):
             info = f"   > Provided efield already good enough!"
             var.cprint(info+f":: emg3d END ::\n", 1)
@@ -416,44 +416,31 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
 
     # Define various l2-norms.
     l2_refe = np.linalg.norm(sfield)  # Reference norm for tolerance.
-    l2_last = np.linalg.norm(residual(grid, model, sfield, efield))
+    _, l2_last = residual(grid, model, sfield, efield)  # Current residual.
     l2_init = l2_last
     l2_prev = l2_last
 
     # If verbose, we keep track on the levels during the first cycle, for QC.
     if var.verb > 2 and var._first_cycle:
 
-        # Initiate _level_all.
-        if level == 0:
-            var._level_all = []
-
         # Store current level.
         var._level_all.append(level)
 
-    # Print initial call info if verbose.
-    if var.verb > 3:
-
-        def print_gs_info(it, level, cycmax, grid, norm, text):
-            """Print info after Gauss-Seidel smoothing steps."""
-            print(f"     {it:2} {level} {cycmax} [{grid.nCx:3}, {grid.nCy:3}, "
-                  f"{grid.nCz:3}]: {norm:.3e} {text}")
-
-        # Print header of smoothing log.
-        if level == 0:
-            print("     it cycmax               error")
-            print("      level [  dimension  ]            info\n")
-            print_gs_info(it, level, cycmax, grid, l2_last, "initial error")
+    # Print initial call info.
+    if level == 0:
+        var.cprint("     it cycmax               error", 3)
+        var.cprint("      level [  dimension  ]            info\n", 3)
+        var.print_gs_info(it, level, cycmax, grid, l2_last, "initial error", 3)
 
     # Initial smoothing (nu_init).
     if level == 0 and var.nu_init > 0:
         # Smooth and re-calculate error.
-        smoothing(grid, model, sfield, efield, var.nu_init, var.lr_dir)
-        l2_last = np.linalg.norm(residual(grid, model, sfield, efield))
+        res, l2_last = smoothing(
+                grid, model, sfield, efield, var.nu_init, var.lr_dir)
 
-        # Print initial smoothing info if verbose.
-        if var.verb > 3:
-            print_gs_info(it, level, cycmax, grid, l2_last,
-                          f"initial smoothing")
+        # Print initial smoothing info.
+        var.print_gs_info(
+                it, level, cycmax, grid, l2_last, f"initial smoothing", 3)
 
     # Start the actual (recursive) multigrid cycle.
     while level == 0 or (level > 0 and it < cycmax):
@@ -467,29 +454,27 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
             # reduces the number of coarsening levels.
 
             # Gauss-Seidel on the coarsest grid.
-            smoothing(grid, model, sfield, efield, var.nu_coarse, var.lr_dir)
+            res, l2_last = smoothing(
+                    grid, model, sfield, efield, var.nu_coarse, var.lr_dir)
 
-            # Print coarsest grid smoothing info if verbose.
-            if var.verb > 3:
-                res = residual(grid, model, sfield, efield)
[…]
