### 2019-05-24 dfa8bb1
Refactoring `restriction`; `MGParameters`; jk-V.
Last jk-refactoring. Done for now.

- move `restr` out of `restriction`;
- move `wx`, `wy`, and `wz` out of `restriction`;
- move most things from `MGParameters` to subfunctions.
