+++ emg3d/solver.py
@@ -33,9 +33,7 @@ from . import utils
 from . import njitted
 
 __all__ = ['solver', 'multigrid', 'smoothing', 'restriction', 'prolongation',
-           'residual', 'krylov', 'MGParameters', 'current_sc_dir',
-           'current_lr_dir', 'print_cycle_info', 'terminate',
-           'RegularGridProlongator']
+           'residual', 'krylov', 'MGParameters', 'RegularGridProlongator']
 
 
 # MAIN USER-FACING FUNCTION
@@ -477,7 +475,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
                 res, l2_last = residual(grid, model, sfield, efield)
 
             # Get sc_dir for this grid.
-            sc_dir = current_sc_dir(var.sc_dir, grid)
+            sc_dir = _current_sc_dir(var.sc_dir, grid)
 
             # (B.2) Restrict grid, model, and fields from fine to coarse grid.
             cgrid, cmodel, csfield, cefield = restriction(
@@ -519,7 +517,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
         else:          # Original grid reached, check termination criteria.
 
             # Print end-of-cycle info.
-            print_cycle_info(var, l2_last, l2_prev, l2_init)
+            _print_cycle_info(var, l2_last, l2_prev, l2_init)
 
             # Adjust semicoarsening and line relaxation if they cycle.
             if var.sc_cycle:
@@ -528,7 +526,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
                 var.lr_dir = next(var.lr_cycle)
 
             # Check if any termination criteria is fulfilled.
-            if terminate(var, l2_last, l2_prev, l2_init, l2_refe, it):
+            if _terminate(var, l2_last, l2_prev, l2_init, l2_refe, it):
                 break
 
             # Set first_cycle to False, to reduce verbosity from now on.
@@ -708,7 +706,7 @@ def smoothing(grid, model, sfield, efield, nu, lr_dir):
            model.eta_z, model.v_mu_r, grid.hx, grid.hy, grid.hz, nu)
 
     # Avoid line relaxation in a direction where there are only two cells.
-    lr_dir = current_lr_dir(lr_dir, grid)
+    lr_dir = _current_lr_dir(lr_dir, grid)
 
     # Calculate and store fields (in-place)
     if lr_dir == 0:             # Standard MG
@@ -787,38 +785,14 @@ def restriction(grid, model, sfield, residual, sc_dir):
         rz = 1
 
     # Calculate distances of coarse grid.
-    ch = [grid.hx, grid.hy, grid.hz]
-    ch[0] = np.diff(grid.vectorNx[::rx])
-    ch[1] = np.diff(grid.vectorNy[::ry])
-    ch[2] = np.diff(grid.vectorNz[::rz])
+    ch = [np.diff(grid.vectorNx[::rx]),
+          np.diff(grid.vectorNy[::ry]),
+          np.diff(grid.vectorNz[::rz])]
 
     # Create new ``TensorMesh``-instance for coarse grid
     cgrid = utils.TensorMesh(ch, grid.x0)
 
     # 2. RESTRICT MODEL
-    def restr(param, sc_dir):
-        """Restrict model parameters."""
-        if sc_dir == 1:    # Only sum the four cells in y-z-plane
-            out = param[:, :-1:2, :-1:2] + param[:, 1::2, :-1:2]
-            out += param[:, :-1:2, 1::2] + param[:, 1::2, 1::2]
-        elif sc_dir == 2:  # Only sum the four cells in x-z-plane
[…]
