### 2019-05-25 ebd2c9d
Cleaning up print statements.
As a result of the refactoring it was more obvious that some things can
be done cleaner/nicer. Probably more might come in the future. Mainly
changes in print statements an the level QC plot.
