+++ emg3d/solver.py
@@ -299,7 +299,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
             linerelaxation=linerelaxation, vnC=grid.vnC, verb=verb, **kwargs
     )
 
-    # Print all parameters if verbose.
+    # Start logging and print all parameters.
     var.cprint(f"\n:: emg3d START :: {var.time.now} ::\n", 1)
     var.cprint(var, 1)
 
@@ -311,18 +311,23 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         # Set flag to return the field.
         do_return = True
     else:
-        # If provided, take the conjugate (see return statement).
+        # If provided, take the conjugate
+        # (see explanation at the end of the solver.solver-routine).
         efield.field = efield.field.conjugate()
+
+        # Set flag to NOT return the field.
+        do_return = False
 
         # If efield is provided, check if it is already sufficiently good.
         _, var.l2 = residual(grid, model, sfield, efield)
         if var.l2 < var.tol*np.linalg.norm(sfield):
-            info = f"   > Provided efield already good enough!"
-            var.cprint(info+f":: emg3d END ::\n", 1)
-            return
 
-        # Set flag to NOT return the field.
-        do_return = False
+            # Switch-off both sslsolver and multigrid.
+            var.sslsolver = None
+            var.cycle = None
+
+            # Start final info.
+            info = f"   > NOTHING DONE (provided efield already good enough)\n"
 
     # Print header for iteration log.
     header = f"   [hh:mm:ss]     {'error':<15}"
@@ -330,14 +335,14 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         header += f"{'solver':<20} "
         if var.cycle:
             header += f"{'MG':<11} l s"
-    else:
-        header += f"{'l2:[last/init, last/prev]':>32} l s"
-    var.cprint(header+"\n", 2)
+        var.cprint(header+"\n", 2)
+    elif var.cycle:
+        var.cprint(header+f"{'l2:[last/init, last/prev]':>32} l s\n", 2)
 
     # Solve the system with...
     if var.sslsolver:  # ... sslsolver.
         krylov(grid, model, sfield, efield, var)
-    else:              # ... multigrid.
+    elif var.cycle:    # ... multigrid.
         multigrid(grid, model, sfield, efield, var)
 
     # Print runtime information.
@@ -345,10 +350,10 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         info = f"   > Solver steps   : {var._ssl_it}\n"
         if var.cycle:
             info += f"   > MG prec. steps : {var.it}\n"
-    else:              # multigrid-specific info.
+    elif var.cycle:    # multigrid-specific info.
         info = f"   > MG cycles      : {var.it}\n"
     info += f"   > Final l2-norm  : {var.l2:.3e}\n\n"  # Final error.
-    info += f":: emg3d END :: {var.time.now} :: "      # END and time.
+    info += f":: emg3d END   :: {var.time.now} :: "    # END and time.
     info += f"runtime = {var.time.runtime}\n"          # Total runtime.
     var.cprint(info, 1)
 
@@ -418,10 +423,8 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
     l2_init = l2_last
     l2_prev = l2_last
 
-    # If verbose, we keep track on the levels during the first cycle, for QC.
-    if var.verb > 2 and var._first_cycle:
-
-        # Store current level.
+    # Keep track on the levels during the first cycle, for QC.
+    if var._first_cycle:
         var._level_all.append(level)
 
     # Print initial call info.
@@ -488,8 +491,8 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
[…]
