### 2019-05-26 66949e1
Only calculate gridE(xyz) when required.
Closes #7.
