### 2019-05-26 482d8f8
Replace how new pts in prolongation are calc.
Allows to remove `gridE*` from utils.TensorMesh; much lighter on memory
usage.

Closes #8.
