### 2019-05-28 2dca8b7
np.subtract inplace operation in solver.residual.
