### 2019-05-28 8377455
np.conjugate inplace operation in solver.solver.
