### 2019-05-28 e9b5181
Only calc. res & l2-norm when absolutely necessary
