+++ emg3d/solver.py
@@ -164,7 +164,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         - 1: Print warnings.
         - 2: Print runtime and information about the method.
         - 3: Print additional information for each MG-cycle.
-        - 4: Print everything.
+        - 4: Print everything (slower due to additional l2-norm calculations).
 
     **kwargs : Optional solver options:
 
@@ -319,7 +319,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         do_return = False
 
         # If efield is provided, check if it is already sufficiently good.
-        _, var.l2 = residual(grid, model, sfield, efield)
+        var.l2 = residual(grid, model, sfield, efield, True)
         if var.l2 < var.tol*np.linalg.norm(sfield):
 
             # Switch-off both sslsolver and multigrid.
@@ -419,7 +419,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
 
     # Define various l2-norms.
     l2_refe = np.linalg.norm(sfield)  # Reference norm for tolerance.
-    _, l2_last = residual(grid, model, sfield, efield)  # Current residual.
+    l2_last = residual(grid, model, sfield, efield, True)  # Current residual.
     l2_init = l2_last
     l2_prev = l2_last
 
@@ -431,17 +431,18 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
     if level == 0:
         var.cprint("     it cycmax               error", 3)
         var.cprint("      level [  dimension  ]            info\n", 3)
-        var.print_gs_info(it, level, cycmax, grid, l2_last, "initial error", 3)
+        if var.verb > 3:
+            _print_gs_info(it, level, cycmax, grid, l2_last, "initial error")
 
     # Initial smoothing (nu_init).
     if level == 0 and var.nu_init > 0:
         # Smooth and re-calculate error.
-        res, l2_last = smoothing(
-                grid, model, sfield, efield, var.nu_init, var.lr_dir)
+        smoothing(grid, model, sfield, efield, var.nu_init, var.lr_dir)
 
         # Print initial smoothing info.
-        var.print_gs_info(
-                it, level, cycmax, grid, l2_last, f"initial smoothing", 3)
+        if var.verb > 3:
+            norm = residual(grid, model, sfield, efield, True)
+            _print_gs_info(it, level, cycmax, grid, norm, "initial smoothing")
 
     # Start the actual (recursive) multigrid cycle.
     while level == 0 or (level > 0 and it < cycmax):
@@ -455,32 +456,30 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
             # reduces the number of coarsening levels.
 
             # Gauss-Seidel on the coarsest grid.
-            res, l2_last = smoothing(
-                    grid, model, sfield, efield, var.nu_coarse, var.lr_dir)
+            smoothing(grid, model, sfield, efield, var.nu_coarse, var.lr_dir)
 
             # Print coarsest grid smoothing info.
-            var.print_gs_info(it, level, cycmax, grid, l2_last,
-                              f"coarsest level", 3)
+            if var.verb > 3:
+                norm = residual(grid, model, sfield, efield, True)
+                _print_gs_info(it, level, cycmax, grid, norm, "coarsest level")
 
         else:                   # (B) Not yet on coarsest grid.
 
             # (B.1) Pre-smoothing (nu_pre).
             if var.nu_pre > 0:
-                res, l2_last = smoothing(
-                        grid, model, sfield, efield, var.nu_pre, var.lr_dir)
[…]
