### 2019-05-30 #1463 @emersodb
In the examples repository, it appears that the folder SEG-EAGE/' is missing from the data/ folder. Thus, when I try to run the 2a_SEG-EAGE_3D-Salt-Model.ipynb notebook, it yields that exception

FileNotFoundError: [Errno 2] No such file or directory: './data/SEG-EAGE/SALTF.ZIP'
