### 2019-05-30 #1463 @prisae
- Data: Correct, you have to download the data. This is described in the notebook just above the cell that failed.
- pymatsolver: Good input! This is strictly speaking a problem with SimPEG, not with emg3d, but I will add a comment in this regard so the user knows why it is not working.
