### 2019-05-30 #1463 @prisae
> You can get the data from the SEG-website or via the direct link. The zip-file is ~1.2 GB big. Unzip the archive, and place the file salt_and_overthrust_models/3-D_Salt_Model/VEL_GRIDS/SALT.ZIP (20.0 MB) into ./data/SEG-EAGE/ (or adjust the path in the following cell).
