### 2019-05-30 #1463 @prisae
No problem at all.

I just added the following sentence under requirements in the notebook `1c_3D_triaxial_SimPEG.ipynb`:
«Note, in order to use the `Pardiso`-solver `pymatsolver` has to be installed via `conda`, not via `pip`!»
Thanks for pointing that out!
