### 2019-05-31 56cb501
Change loop-order in gauss_seidel_y
Change loop-order from x-z-y to z-x-y in `gauss_seidel_y` (Gauss-Seidel smoothing with line relaxation in y-direction). Hence reversed lexicographical order. This results in a significant speed-up, as x is the fastest changing axis (F-ordered array).
