### 2019-05-31 69996c3
Simplify get_source_field
