### 2019-05-31 36891f1
Remove res from utils.Model.
`utils.Model.res(_x,_y,_z)` is just for convenience, it is not used by
multigrid, as calculation is carried out with eta. Remove res to save
some memory. Resistivity can still be set and obtained, but then it is
calculated on the fly from eta.
