+++ emg3d/utils.py
@@ -756,63 +756,6 @@ class Model:
     def __init__(self, grid, res_x=1., res_y=None, res_z=None, freq=1.):
         """Initiate a new resistivity model."""
 
-        # Initiate resistivity, eta, and mu_r
-        if res_y is None and res_z is None:   # Isotropic.
-            self.case = 0
-            self.__res = np.empty(grid.nC, dtype=float)
-            self.__eta = np.empty(grid.nC, dtype=complex)
-        elif res_y is None or res_z is None:  # VTI or HTI.
-            if res_z is None:
-                self.case = 1
-            else:
-                self.case = 2
-            self.__res = np.empty(2*grid.nC, dtype=float)
-            self.__eta = np.empty(2*grid.nC, dtype=complex)
-        else:                                 # Tri-axial anisotropy.
-            self.case = 3
-            self.__res = np.empty(3*grid.nC, dtype=float)
-            self.__eta = np.empty(3*grid.nC, dtype=complex)
-
-        # Set x-directed resistivity
-        if isinstance(res_x, (float, int)):
-            self.__res[:grid.nC] = res_x*np.ones(grid.nC)
-        else:
-            if res_x.shape[0] != grid.nC:
-                print("* ERROR   :: Provided input res_x has shape "
-                      f"{res_x.shape}; but shape {grid.nC} is required.")
-                raise ValueError("Wrong Shape")
-            self.__res[:grid.nC] = res_x
-
-        # Set y-directed resistivity
-        if res_y is None:
-            pass
-        elif isinstance(res_y, (float, int)):
-            self.__res[grid.nC:2*grid.nC] = res_y*np.ones(grid.nC)
-        else:
-            if res_y.shape[0] != grid.nC:
-                print("* ERROR   :: Provided input res_y has shape "
-                      f"{res_y.shape}; but shape {grid.nC} is required.")
-                raise ValueError("Wrong Shape")
-            self.__res[grid.nC:2*grid.nC] = res_y
-
-        # Set z-directed resistivity
-        if res_z is None:
-            pass
-        elif isinstance(res_z, (float, int)):
-            if res_y is None:
-                self.__res[grid.nC:] = res_z*np.ones(grid.nC)
-            else:
-                self.__res[2*grid.nC:] = res_z*np.ones(grid.nC)
-        else:
-            if res_z.shape[0] != grid.nC:
-                print("* ERROR   :: Provided input res_z has shape "
-                      f"{res_z.shape}; but shape {grid.nC} is required.")
-                raise ValueError("Wrong Shape")
-            if res_y is None:
-                self.__res[grid.nC:] = res_z
-            else:
-                self.__res[2*grid.nC:] = res_z
-
         # Store frequency
         self.freq = freq
 
@@ -821,8 +764,62 @@ class Model:
         self.vnC = grid.vnC
         self.vol = grid.vol
 
-        # Update eta with given resistivities
-        self.update_eta
+        # Initiate resistivity, eta, and mu_r
+        if res_y is None and res_z is None:   # Isotropic.
+            self.case = 0
+            self.__eta = np.empty(grid.nC, dtype=complex)
+        elif res_y is None or res_z is None:  # VTI or HTI.
+            if res_z is None:
+                self.case = 1
+            else:
+                self.case = 2
+            self.__eta = np.empty(2*grid.nC, dtype=complex)
+        else:                                 # Tri-axial anisotropy.
+            self.case = 3
+            self.__eta = np.empty(3*grid.nC, dtype=complex)
+
+        # Set x-directed resistivity
+        if isinstance(res_x, (float, int)):
[…]
