### 2019-05-31 478ce2a
Move total residual calc. to amat_x
