### 2019-05-31 2738dda
Re-write utils.Model-class.
- Mostly undo/redo 36891; use resistivity only instead of eta only.
- Remove ability to set eta. Just used internally anyway.
- Use a simple dummy-class inside solver.restriction; only have
eta_(x,y,z) and v_mu_r and case.
