### 2019-06-01 4235b72
Further simplification of Model and TensorMesh.
