+++ emg3d/utils.py
@@ -714,14 +714,6 @@ class TensorMesh:
         self.vnE = np.array([self.nEx, self.nEy, self.nEz], dtype=int)
         self.nE = int(self.vnE.sum())
 
-    @property
-    def vol(self):
-        """Construct cell volumes of the 3D model as 1d array."""
-        if getattr(self, '_vol', None) is None:
-            xy = np.outer(self.hx, self.hy).flatten(order='F')
-            self._vol = np.outer(xy, self.hz).flatten(order='F')
-        return self._vol
-
 
 # MODEL AND FIELD CLASSES
 class Model:
@@ -762,61 +754,67 @@ class Model:
         # Store required info from grid
         self.nC = grid.nC
         self.vnC = grid.vnC
-        self.vol = grid.vol
 
-        # Initiate resistivity and mu_r
+        # Construct cell volumes of the 3D model as 1d array.
+        vol = np.outer(np.outer(grid.hx, grid.hy).ravel('F'), grid.hz)
+        self.__vol = vol.ravel('F').reshape(self.vnC, order='F')
+
+        # Get case
         if res_y is None and res_z is None:   # Isotropic.
             self.case = 0
-            self.__res = np.empty(grid.nC, dtype=float)
         elif res_y is None or res_z is None:  # VTI or HTI.
             if res_z is None:
                 self.case = 1
             else:
                 self.case = 2
-            self.__res = np.empty(2*grid.nC, dtype=float)
         else:                                 # Tri-axial anisotropy.
             self.case = 3
-            self.__res = np.empty(3*grid.nC, dtype=float)
 
         # Set x-directed resistivity
         if isinstance(res_x, (float, int)):
-            self.__res[:grid.nC] = res_x*np.ones(grid.nC)
+            self.__res_x = res_x*np.ones(self.vnC, order='F')
         else:
-            if res_x.shape[0] != grid.nC:
+            if res_x.size != self.nC:
                 print("* ERROR   :: Provided input res_x has shape "
-                      f"{res_x.shape}; but shape {grid.nC} is required.")
+                      f"{res_x.shape}; but shape {self.vnC} or {self.nC} is "
+                      f"required.")
                 raise ValueError("Wrong Shape")
-            self.__res[:grid.nC] = res_x
+            elif res_x.shape[0] == self.vnC[0]:
+                self.__res_x = res_x
+            else:
+                self.__res_x = res_x.reshape(self.vnC, order='F')
 
         # Set y-directed resistivity
         if res_y is None:
             pass
         elif isinstance(res_y, (float, int)):
-            self.__res[grid.nC:2*grid.nC] = res_y*np.ones(grid.nC)
+            self.__res_y = res_y*np.ones(self.vnC, order='F')
         else:
-            if res_y.shape[0] != grid.nC:
+            if res_y.size != self.nC:
                 print("* ERROR   :: Provided input res_y has shape "
-                      f"{res_y.shape}; but shape {grid.nC} is required.")
+                      f"{res_y.shape}; but shape {self.vnC} or {self.nC} is "
+                      f"required.")
                 raise ValueError("Wrong Shape")
-            self.__res[grid.nC:2*grid.nC] = res_y
+            elif res_y.shape[0] == self.vnC[0]:
+                self.__res_y = res_y
+            else:
+                self.__res_y = res_y.reshape(self.vnC, order='F')
 
         # Set z-directed resistivity
         if res_z is None:
             pass
         elif isinstance(res_z, (float, int)):
-            if res_y is None:
-                self.__res[grid.nC:] = res_z*np.ones(grid.nC)
-            else:
[…]
