### 2019-06-03 0a40d52
Another try at simplifying utils.Model
