+++ emg3d/utils.py
@@ -719,12 +719,9 @@ class TensorMesh:
 class Model:
     r"""Create a resistivity model.
 
-    For now, the model grid equals the calculation grid. However, this should
-    change in the future, and functions to go from one to the other and back
-    have to be implemented.
-
-    Relative electric permeability :math:`\varepsilon_r` and relative
-    magnetic permittivity :math:`\mu_r` are both fixed at 1.
+    Class to provide model parameters (x-, y-, and z-directed resistivities)
+    to the solver. Relative electric permeability :math:`\varepsilon_r` and
+    relative magnetic permittivity :math:`\mu_r` are both fixed at 1.
 
     Parameters
     ----------
@@ -732,94 +729,82 @@ class Model:
         Grid on which to apply model.
 
     res_x, res_y, res_z : float or ndarray; default to 1.
-        Resistivity in x-, y-, and z-directions. If ndarray, the must have
-        length of grid.nC.
+        Resistivity in x-, y-, and z-directions. If ndarray, they must have the
+        shape of grid.vnC (F-ordered) or grid.nC.
 
     freq : float
         Frequency.
 
     """
 
-    # Constants
-    c = 299792458              # Speed of light m/s
-    mu_0 = 4e-7*np.pi          # Magn. permeability of free space [H/m]
-    epsilon_0 = 1./(mu_0*c*c)  # Elec. permittivity of free space [F/m]
-
     def __init__(self, grid, res_x=1., res_y=None, res_z=None, freq=1.):
         """Initiate a new resistivity model."""
 
-        # Store frequency
+        # Store frequency.
         self.freq = freq
 
-        # Store required info from grid
+        # Store required info from grid.
         self.nC = grid.nC
         self.vnC = grid.vnC
 
-        # Construct cell volumes of the 3D model as 1d array.
-        vol = np.outer(np.outer(grid.hx, grid.hy).ravel('F'), grid.hz)
-        self.__vol = vol.ravel('F').reshape(self.vnC, order='F')
+        # Construct cell volumes of the 3D model as 1D array.
+        if hasattr(grid, 'vol'):  # If discretize-grid, take it from there.
+            self.__vol = grid.vol.reshape(self.vnC, order='F')
+        else:                     # Calculate it otherwise.
+            vol = np.outer(np.outer(grid.hx, grid.hy).ravel('F'), grid.hz)
+            self.__vol = vol.ravel('F').reshape(self.vnC, order='F')
 
-        # Get case
-        if res_y is None and res_z is None:   # Isotropic.
+        # Check case.
+        if res_y is None and res_z is None:   # Isotropic (0).
             self.case = 0
-        elif res_y is None or res_z is None:  # VTI or HTI.
+        elif res_y is None or res_z is None:  # HTI (1) or VTI (2).
             if res_z is None:
                 self.case = 1
             else:
                 self.case = 2
-        else:                                 # Tri-axial anisotropy.
+        else:                                 # Tri-axial anisotropy (3).
             self.case = 3
 
-        # Set x-directed resistivity
+        # Initiate x-directed resistivity.
         if isinstance(res_x, (float, int)):
-            self.__res_x = res_x*np.ones(self.vnC, order='F')
+            self.__res_x = res_x*np.ones(self.vnC)
+        elif np.all(res_x.shape == self.vnC) and res_x.ndim == 3:
+            self.__res_x = res_x
+        elif res_x.size == self.nC and res_x.ndim == 1:
+            self.__res_x = res_x.reshape(self.vnC, order='F')
         else:
-            if res_x.size != self.nC:
-                print("* ERROR   :: Provided input res_x has shape "
[…]
