### 2019-06-03 aae1a39
Clean-up timing
