### 2019-06-03 85bd4a3
New doc re CPU & RAM
