+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 Memory and speed improvements:
 
+- Add *CPU & RAM*-page to documentation.
 - Change loop-order from x-z-y to z-x-y in Gauss-Seidel smoothing with line
   relaxation in y-direction. Hence reversed lexicographical order. This results
   in a significant speed-up, as x is the fastest changing axis.
