### 2019-06-03 #1463 @prisae
`v0.6.2` is out. All these new versions are changes under the hood to improve CPU and RAM usage, nothing changes for the user-facing functionality. I do not expect further new versions until the JOSS review is finished (if it is not taking too long).
