+++ docs/cpu_and_ram.rst
@@ -1,17 +1,29 @@
 CPU & RAM
 #########
 
-Some notes regarding memory and runtime requirements; amongst others things
-that have been tried, and things that could be tried.
+The multigrid method is attractive because it shows optimal scaling for both
+runtime and memory consumption. In the following are a few notes regarding
+memory and runtime requirements. It also contains information about what has
+been tried and what still could be tried in order to improve the current code.
 
 
 Runtime
 -------
 
-The costliest functions (RAM- and CPU-wise) are already written in ``numba``.
-There might be an additional gain by making :class:`emg3d.utils.TensorMesh`,
-:class:`emg3d.utils.Model`, and :class:`emg3d.utils.Field` instances jitted
-classes. Keep an eye on the progress of jitted classes within numba.
+The `examples-gallery <https://github.com/empymod/emg3d-examples>`_ contains a
+notebook to do some testing with regards to runtime, see
+`4b_Runtime.ipynb
+<https://github.com/empymod/emg3d-examples/blob/master/4b_Runtime.ipynb>`_.
+An example output of that notebook is shown in :numref:`Figure %s <runtime>`.
+
+.. figure:: _static/runtime.svg
+   :scale: 80 %
+   :align: center
+   :alt: Runtime
+   :name: runtime
+
+   Runtime as a function of cell size, which shows nicely the linear scaling
+   of multigrid solvers.
 
 The costliest functions (for big models) are:
 
@@ -35,11 +47,25 @@ The rest can be ignored. For small models, the percentage of ``smoothing`` goes
 down and of ``prolongation`` and ``restriction`` go up. But then the modeller
 is fast anyway.
 
-**Therefore, any serious attempt to improve the speed would have to tackle the
-smoothing itself.**
+:func:`emg3d.njitted.gauss_seidel` and :func:`emg3d.njitted.amat_x` are written
+in ``numba``; jitting :class:`emg3d.solver.RegularGridProlongator` turned out
+to not improve things, and many functions used in the restriction are jitted
+too. The costliest functions (RAM- and CPU-wise) are therefore already written
+in ``numba``.
+
+**Any serious attempt to improve the speed will have to tackle the smoothing
+itself.**
 
 
-**Tried things**
+**Things which could be tried**
+
+- Not much has been tested with the ``numba``-options ``parallel``; ``prange``;
+  and ``nogil``.
+- There might be an additional gain by making :class:`emg3d.utils.TensorMesh`,
+  :class:`emg3d.utils.Model`, and :class:`emg3d.utils.Field` instances jitted
+  classes.
+
+**Things which have been tried**
 
 - One important aspect of the smoothing part is the memory layout.
   :func:`emg3d.njitted.gauss_seidel` and :func:`emg3d.njitted.gauss_seidel_x`
@@ -49,26 +75,20 @@ smoothing itself.**
   But copying the arrays to C-order and afterwards back is costlier in most
   cases for both CPU and RAM. The one possible and therefore implemented
   solution was to swap the loop-order in :func:`emg3d.njitted.gauss_seidel_y`.
-- Restriction/prolongation information could be saved in a dictionary instead
-  of recalculating it every time. Turns out to be not worth the trouble.
+- Restriction and prolongation information could be saved in a dictionary
+  instead of recalculating it every time. Turns out to be not worth the
+  trouble.
 - Rewrite :class:`emg3d.RegularGridInterpolator` as jitted function, but the
   iterator approach seems to be better for large grids.
-
-**Not tried**
-
-- Not much has been tested with the ``numba``-options ``parallel``; ``prange``;
-  and ``nogil``.
 
 
 Memory
 ------
 
-Estimating the memory requirement is never an easy task. Here a few attempts.
-
 Most of the memory requirement comes from storing the data itself, mainly the
[…]
