### 2019-06-05 #1463 @prisae
That is a good and timely question @emersodb. Since version 0.6.2 (so two days ago) I had a _CPU & RAM_ section in the docs, and a notebook to estimate memory. Given your question I expanded this section and also added a notebook to estimate runtime.

- Docs: https://emg3d.readthedocs.io/en/latest/cpu_and_ram.html
- Notebooks: `4a_RAM-requirements.ipynb` and `4b_Runtime.ipynb` on https://github.com/empymod/emg3d-examples

Figures

![runtime](https://user-images.githubusercontent.com/8020943/58943594-5699e280-8780-11e9-9074-862d2f47b641.png)

![RAM-Usage](https://user-images.githubusercontent.com/8020943/58943611-60bbe100-8780-11e9-8293-ac1a9eeefea8.png)
