### 2019-06-06 9ea5e8d
Isotropic magnetic permeability.
