### 2019-06-06 115965e
Expand `utils.Field` to work with pickle/shelve.
