### 2019-06-06 6dda67e
Extend writing/loading data
