### 2019-06-07 f8945c5
Add magnetic field and receiver routines.
- utils.get_h_field: Obtain magnetic field from electric field.
- utils.get_receiver: Simple interpolation wrapper to get field at
  receiver locations.
