+++ emg3d/utils.py
@@ -35,6 +35,7 @@ import numpy as np
 import multiprocessing
 from scipy import optimize
 from timeit import default_timer
+from scipy import interpolate as sint
 from datetime import datetime, timedelta
 
 import emg3d  # emg3d for Versions
@@ -60,9 +61,9 @@ else:
     mklinfo = False
 
 
-__all__ = ['get_domain', 'get_stretched_h', 'get_hx', 'get_source_field',
-           'Model', 'Field', 'TensorMesh', 'Time', 'data_write', 'data_read',
-           'Versions']
+__all__ = ['Model', 'Field', 'get_domain', 'get_stretched_h', 'get_hx',
+           'get_source_field', 'get_receiver', 'get_h_field', 'TensorMesh',
+           'Time', 'data_write', 'data_read', 'Versions']
 
 
 # CONSTANTS
@@ -714,7 +715,7 @@ class TensorMesh:
         self.nE = int(self.vnE.sum())
 
 
-# MODEL AND FIELD CLASSES
+# RELATED TO MODELS AND FIELDS
 class Model:
     r"""Create a resistivity model.
 
@@ -1065,6 +1066,151 @@ class Field(np.ndarray):
         self.fz[-1, :, :] = 0.+0.j
         self.fz[:, 0, :] = 0.+0.j
         self.fz[:, -1, :] = 0.+0.j
+
+
+def get_h_field(grid, model, field):
+    r"""Return magnetic field corresponding to provided electric field.
+
+    Retrieve the magnetic field :math:`\mathbf{H}` from the electric field
+    :math:`\mathbf{H}`  using Farady's law, given by
+
+    .. math::
+
+        \nabla \times \mathbf{E} = \rm{i}\omega\mu\mathbf{H} .
+
+    Note that the magnetic field in x-direction is defined in the center of the
+    face defined by the electric field in y- and z-directions, and similar for
+    the other field directions. This means that the provided electric field and
+    the returned magnetic field have different dimensions::
+
+       E-field:  x: [grid.vectorCCx,  grid.vectorNy,  grid.vectorNz]
+                 y: [ grid.vectorNx, grid.vectorCCy,  grid.vectorNz]
+                 z: [ grid.vectorNx,  grid.vectorNy, grid.vectorCCz]
+
+       H-field:  x: [ grid.vectorNx, grid.vectorCCy, grid.vectorCCz]
+                 y: [grid.vectorCCx,  grid.vectorNy, grid.vectorCCz]
+                 z: [grid.vectorCCx, grid.vectorCCy,  grid.vectorNz]
+
+
+    Parameters
+    ----------
+    grid : TensorMesh
+        Model grid; ``emg3d.utils.TensorMesh`` instance.
+
+    model : Model
+        Model; ``emg3d.utils.Model`` instance.
+
+    field : Field
+        Electric field; ``emg3d.utils.Field`` instance.
+
+
+    Returns
+    -------
+    hfield : Field
+        Magnetic field; ``emg3d.utils.Field`` instance.
+
+    """
+    # Define the widths of the dual grid.
+    dx = (np.r_[0., grid.hx] + np.r_[grid.hx, 0.])/2.
+    dy = (np.r_[0., grid.hy] + np.r_[grid.hy, 0.])/2.
+    dz = (np.r_[0., grid.hz] + np.r_[grid.hz, 0.])/2.
+
+    # If relative magnetic permeability is not one, we have to take the volume
[…]
