+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Changelog
   - ``utils.get_h_field``: Small routine to calculate the magnetic field from
     the electric field using Faraday's law.
   - ``utils.get_receiver``: Small wrapper to interpolate a field at receiver
-    positions.
+    positions. Added 3D spline interpolation; is the new the default.
 
 - Re-implemented the possibility to define isotropic magnetic permeabilities in
   ``utils.Model``. Magnetic permeability is not tri-axially included in the
