### 2019-06-12 0bc32d0
Improve verbosity of write_data
