### 2019-06-12 #1463 @emersodb
At this time, I believe that my review is complete and would recommend publication of the emg3d submission.
