### 2019-06-13 130724b
Docstrings
