### 2019-06-13 d470b24
Improve data_write and data_read
