### 2019-06-13 #1463 @prisae
:question: We want to cite this article. I assume this review is not done within a week or two, so what is the preferred way of citing a JOSS article under review? Just the usual "submitted to JOSS"?

In the proof, the article has already a DOI. I assume this will remain the same and is not a dummy-DOI for the time being, correct?

Adjusted version from the proof:
> Werthmüller et al., (2019). emg3d: A multigrid solver for 3D electromagnetic diffusion. Submitted to the Journal of Open Source Software, 4(37), 1463; https://doi.org/10.21105/joss.01463.
