### 2019-06-13 #1463 @jedbrown
That DOI hasn't been created yet, but that is the DOI that will be created upon acceptance. Do you intend to cite this JOSS paper from a new submission or from a camera-ready accepted manuscript?
