### 2019-06-13 #1463 @prisae
That is fine, then it is as I understood it. It hasn't been created, but it will be created exactly like this (`10.21105/joss.01463`) once it is accepted and in its final form.

It is a new submission; it will still take some time until this will be in print or published. I hope/expect that by then this review will be finished and the DOI created.
