### 2019-06-13 98c021d
Fix test_Time() fail.
Closes #13.
