### 2019-06-13 empymod/emg3d#14 @lukeolson
Basic Example
I could use some help setting up a basic problem on a unit cube (or square).  I've been able to run the following but trying a forcing function of say all ones has been difficult.

```
h0 = 0.1 * np.ones(10)
grid = emg3d.utils.TensorMesh((h0,h0,h0), np.zeros(3))
model = emg3d.utils.Model(grid)
sfield = emg3d.utils.Field(grid)
efield = emg3d.solver.solver(grid, model, sfield)
```

A couple questions:
1. with model, it seems that the permitivity and permeability are both 1 (with 1/\sigma being a parameter).  Is it possible to set all of \eta, \mu_r, \omega, \mu_0 through this interface or is this intended to be a simpler model?
2. With the source field it appears that theses are edge values, but I can't seem to get a Field(fx,fy,fz) initiated.
