### 2019-06-13 empymod/emg3d#14 @prisae
`emg3d.utils.TensorMesh` is VERY basic; it is mainly used by the solver itself, but it can also be used to set up a grid. However, for fancier/easier set-ups it is easier to use `discretize.TensorMesh` (`pip install discretize`).  http://discretize.simpeg.xyz
