### 2019-06-13 empymod/emg3d#14 @prisae
`emg3d.utils.Model` - Model builder:

- In the current stable version, the magnetic permeability is fixed to 1. On master (and hence from v0.6.3 onwards), you can set the magnetic permeability by providing `emg3d.utils.Model(... mu_r)`. Only isotropic though.

- You cannot set the `eta_(x,y,z)`, but all of the resistivities, by providing `emg3d.utils.Model(...res_x, res_y, res_z)`. Now, `res_(x,y,z)` can be floats, or arrays of dimension `(nx, ny, nz)`=`grid.vnC`, or even an array of length `nx*ny*nz`=`grid.nC` (Fortran-ordered).

- You can provide the frequency with the `freq` parameter, which then will set `omega` (2*pi*freq).

- `mu_0` cannot be set, it is a physical constant.

Do you think it needs more in the documentation?
https://emg3d.readthedocs.io/en/latest/code.html#emg3d.utils.Model

Example:
```
h0 = 0.1 * np.ones(10)
grid = emg3d.utils.TensorMesh((h0,h0,h0), np.zeros(3))

res_x = 1
res_z = np.arange(grid.nC)+1
freq = 10
model = emg3d.utils.Model(grid, res_x=res_x, res_z=res_z, freq=freq)
```

This sets-up a model with frequency = 10 Hz. x-directed resistivity is 1 Ohm m, z-directed resistivity is increasing for each cell from 1 to number of cells Ohm m. As I did not provide `res_y` it assumes a VTI-model, hence same resistivity in x- and y-direction.

On top of the API-docs, the example might give some hints: https://github.com/empymod/emg3d-examples.

If you think I need to document that better I'll be happy to do so. Maybe add a getting-started notebook in the examples. However, I am off to holiday for 10 days, so I'll get back to it then.

Also, an easy getting-started example with some comments is given in the docs: https://emg3d.readthedocs.io/en/stable/usage.html
