### 2019-06-13 empymod/emg3d#14 @prisae
### 3. Source field

This `sfield = emg3d.utils.Field(grid)` actually gives you a `Field(fx, fy, fz)`; try `sfield.fx`, `sfield.fy`, and `sfield.fz`.

However, to set-up a source field you need the routine

`sfield = emg3d.utils.get_source_field(grid, src=[0, 0, 0, 0, 0], freq=freq)`

or you can, as you did, initiate an empty field and then put your source wherever you want in that.
