### 2019-06-13 empymod/emg3d#14 @prisae
Let me know if these helps, and if you think I need more on top of:

- https://emg3d.readthedocs.io/en/stable/usage.html
- https://github.com/empymod/emg3d-examples
- https://emg3d.readthedocs.io/en/latest/code.html#emg3d.utils.Model
- https://emg3d.readthedocs.io/en/latest/code.html#emg3d.utils.Field
- https://emg3d.readthedocs.io/en/latest/code.html#emg3d.utils.get_source_field
- https://emg3d.readthedocs.io/en/latest/code.html#emg3d.utils.TensorMesh
