### 2019-06-13 empymod/emg3d#14 @prisae
The only thing you would have to change in your example is

`sfield = emg3d.utils.get_source_field(grid, src=[0.5, 0.5, 0.5, 0, 0], freq=1)`

This sets up a x-directed (azimuth=0, dip=0) source at location (0.5, 0.5, 0.5); frequency=1 (same as you used to set-up the model).

Then

```
In [9]: efield = emg3d.solver.solver(grid, model, sfield)

:: emg3d START :: 22:03:58 ::

   MG-cycle       : 'F'                 sslsolver : False
   semicoarsening : False [0]           tol       : 1e-06
   linerelaxation : False [0]           maxit     : 50
   nu_{i,1,c,2}   : 0, 2, 1, 2          verb      : 2
   Original grid  :  10 x  10 x  10     => 1,000 cells
   Coarsest grid  :   5 x   5 x   5     => 125 cells
   Coarsest level :   1 ;   1 ;   1

   > CONVERGED
   > MG cycles      : 16
   > Final l2-norm  : 5.012e-12

:: emg3d END   :: 22:03:58 :: runtime = 0:00:00
```

You get some additional goodies if you initiate your grid with `discretize` instead of `emg3d`:

`grid = discretize.TensorMesh((h0,h0,h0), np.zeros(3))`

Then you can plot your model and the resulting field:

`grid.plot_3d_slicer(efield.fx.ravel('F'), view='abs', vType='Ex')`

which results in

![index](https://user-images.githubusercontent.com/8020943/59464341-77b99d80-8e28-11e9-8f66-4ed95e991a6f.png)

which is an interactive figure, depending on your backend, so you can slice through it.

Similarly, you could plot the model parameters:

`grid.plot_3d_slicer(model.res_x)`

Now, these are `discretize`-features, not `emg3d`-features, but they come in handy.
