### 2019-06-13 empymod/emg3d#14 @prisae
![Peek 2019-06-13 22-17](https://user-images.githubusercontent.com/8020943/59464557-0dedc380-8e29-11e9-881e-a1d98b2c383c.gif)

Needs `%matplotlib notebook` in a notebook, in IPython it should work out of the box.
