### 2019-06-25 empymod/emg3d#14 @prisae
Hi @lukeolson. I am back from holiday and catching up. Did my comments answer your initial questions?
