### 2019-06-26 955b925
Adjust intersphinx-link for discretize.
Resolves #15.
