+++ docs/conf.py
@@ -25,7 +25,7 @@ todo_include_todos = True
 intersphinx_mapping = {
     "numpy": ("https://docs.scipy.org/doc/numpy/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/reference", None),
-    "discretize": ("https://discretize.readthedocs.io/en/stable", None),
+    "discretize": ("https://discretize.simpeg.xyz/en/master", None),
 }
 
 # ==== 2. General Settings ====
