### 2019-06-26 7390509
DOC and typos; bugfix in utils.get_hx
