### 2019-06-27 empymod/emg3d#14 @lukeolson
Yep, this helps.  One suggestion is a to provide a very basic example with some of these comments.  I used
```
import emg3d
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

h0 = 0.1 * np.ones(10)
grid = emg3d.utils.TensorMesh((h0,h0,h0), np.zeros(3))

res_x = 1
res_z = np.arange(grid.nC)+1
freq = 10
model = emg3d.utils.Model(grid, res_x=res_x, res_z=res_z, freq=freq)
sfield = emg3d.utils.get_source_field(grid, src=[0.5, 0.5, 0.5, 0, 0], freq=1)

efield = emg3d.solver.solver(grid, model, sfield)
```

One question from this: why is the grid 10x10x10 in this case, but the field variables (e.g. Ex) 10x11x11?
