### 2019-06-27 #1463 @lukeolson
I believe everything is done on my checklist.  One suggestion was to add a basic example of usage: https://github.com/empymod/emg3d/issues/14
