### 2019-06-27 #1463 @prisae
Thanks @lukeolson for your review!

Re basic example: Is the one basic example that is already in the docs under the *Getting started* section not enough? I think it is pretty similar to yours:
https://emg3d.readthedocs.io/en/stable/usage.html#example
