### 2019-06-27 empymod/emg3d#14 @prisae
@lukeolson:

With regards to your question:

The grid is 10 x 10 x 10 cells. It is a Yee-grid, so the electric fields are located on the edges: See https://emg3d.readthedocs.io/en/stable/theory.html#muld06-fig1

That means that
- Ex has dimension 10x11x11  (cells x edges x edges)
- Ey has dimension 11x10x11  (edges x cells x edges)
- Ez has dimension 11x11x10  (edges x edges x cells)

Does that make sense?

I close this issue then. There is a very similar basic example given in https://emg3d.readthedocs.io/en/stable/usage.html#example
