### 2019-06-27 #1463 @lukeolson
Yes, that's probably fine.  I agree my small example is nearly the same in the end.
