### 2019-06-27 empymod/emg3d#14 @lukeolson
Ah -- a Yee-grid.  Yep that makes sense.
