### 2019-06-27 #1463 @prisae
One more question if you don't mind @lukeolson: I saw from your website that you are part of [PyAMG](http://pyamg.github.io/), so I assume that you have experience in multigrid methods. Do you have any feedback from the multigrid-viewpoint in particular?
