### 2019-06-28 empymod/emg3d#14 @prisae
I expanded a bit on my minimum example, and also created a notebook out of it, so it should be easier to get started:
https://github.com/empymod/emg3d-examples/blob/master/0a_Minimum_working_example.ipynb
