### 2019-06-28 c44dad5
Improve simple example in docs
