+++ emg3d/utils.py
@@ -24,41 +24,12 @@ Utility functions for the multigrid solver.
 
 
 import os
-import sys
-import time
-import numba
-import scipy
+import scooby
 import shelve
-import textwrap
-import platform
 import numpy as np
-import multiprocessing
 from timeit import default_timer
 from datetime import datetime, timedelta
 from scipy import optimize, interpolate, ndimage
-
-import emg3d  # emg3d for Versions
-
-# Optional modules
-try:
-    import IPython
-except ImportError:
-    IPython = False
-try:
-    import matplotlib
-except ImportError:
-    matplotlib = False
-try:
-    import mkl
-except ImportError:
-    mkl = False
-
-# Get mkl info, if available
-if mkl:
-    mklinfo = mkl.get_version_string()
-else:
-    mklinfo = False
-
 
 __all__ = ['Model', 'Field', 'get_domain', 'get_stretched_h', 'get_hx',
            'get_source_field', 'get_receiver', 'get_h_field', 'TensorMesh',
@@ -1435,183 +1406,16 @@ def data_read(fname, keys=None, path="data"):
 
 
 # OTHER
-class Versions:
-    r"""Print date, time, and version information.
+class Versions(scooby.Report):
 
-    Print date, time, and package version information in any environment
-    (Jupyter notebook, IPython console, Python console, QT console), either as
-    html-table (notebook) or as plain text (anywhere).
+    def __init__(self, add_pckg=None, ncol=4, **kwargs):
+        """Initiate a scooby.Report instance."""
 
-    Always shown are the OS, number of CPU(s), ``numpy``, ``scipy``,
-    ``emg3d``, ``numba``, ``sys.version``, and time/date.
+        # Mandatory packages.
+        core = ['numpy', 'scipy', 'numba', 'emg3d']
 
-    Additionally shown are, if they can be imported, ``IPython`` and
-    ``matplotlib``. It also shows MKL information, if available.
+        # Optional packages.
+        optional = ['IPython', 'matplotlib']
 
-    All modules provided in ``add_pckg`` are also shown. They have to be
-    imported before ``versions`` is called.
-
-    This script was heavily inspired by:
-
-        - ipynbtools.py from qutip https://github.com/qutip
-        - watermark.py from https://github.com/rasbt/watermark
-
-
-    Parameters
-    ----------
-    add_pckg : packages, optional
-        Package or list of packages to add to output information (must be
-        imported beforehand).
-
-    ncol : int, optional
-        Number of package-columns in html table; only has effect if
-        ``mode='HTML'`` or ``mode='html'``. Defaults to 3.
-
-
-    Examples
-    --------
-    >>> import pytest
-    >>> import dateutil
-    >>> from emg3d import Versions
-    >>> Versions()                            # Default values
-    >>> Versions(pytest)                      # Provide additional package
-    >>> Versions([pytest, dateutil], ncol=5)  # Set nr of columns
-
-    """
-
-    def __init__(self, add_pckg=None, ncol=4):
-        """Initiate and add packages and number of columns to self."""
-        self.add_pckg = self._get_packages(add_pckg)
-        self.ncol = int(ncol)
-
-    def __repr__(self):
-        r"""Plain-text version information."""
-
-        # Width for text-version
-        n = 54
-        text = '\n' + n*'-' + '\n'
-
-        # Date and time info as title
-        text += time.strftime('  %a %b %d %H:%M:%S %Y %Z\n\n')
-
-        # OS and CPUs
-        text += '{:>15}'.format(platform.system())+' : OS\n'
-        text += '{:>15}'.format(multiprocessing.cpu_count())+' : CPU(s)\n'
-
-        # Loop over packages
-        for pckg in self.add_pckg:
[…]
