+++ .travis.yml
@@ -28,7 +28,7 @@ install:
   # Install and activate environment, install packages
   - conda create -q -n test-environment -c $CHAN python=$PYTHON numpy scipy pytest pytest-cov $PCKGS
   - source activate test-environment
-  - pip install coveralls pytest-flake8 $INST
+  - pip install coveralls pytest-flake8 scooby $INST
   - python setup.py install
 
 
