### 2019-07-01 #1463 @jedbrown
@akelbert Is there anything we can do to facilitate your review? Please note that it is fine to review from the perspective of a potential user for joint inversion; other reviewers have covered different perspectives and you need not spend time in the overlap if time is limited. Thanks.
