### 2019-07-01 #1463 @prisae
> from the perspective of a potential user for joint inversion

There is no (joint) inversion in `emg3d`, I assume you simply meant "from the perspective of a potential user" or "from the perspective of a potential user for modelling".
