### 2019-07-01 fa4aa48
Simplify tests
