+++ setup.py
@@ -25,6 +25,6 @@ setup(
         'numpy>=1.15.0',
         'scipy>=1.1.0',
         'numba>=0.40.0',
-        'scooby>0.2.0'
+        'scooby>=0.3.0'
     ],
 )
