+++ requirements.txt
@@ -5,7 +5,7 @@ sphinx_rtd_theme
 numpy >= 1.15.0
 scipy >= 1.1.0
 numba >= 0.40.0
-scooby > 0.2
+scooby >= 0.3.0
 
 # sphinx fixed below 2.0
 # https://github.com/numpy/numpydoc/issues/215
