+++ emg3d/utils.py
@@ -24,41 +24,12 @@ Utility functions for the multigrid solver.
 
 
 import os
-import sys
-import time
-import numba
-import scipy
+import scooby
 import shelve
-import textwrap
-import platform
 import numpy as np
-import multiprocessing
 from timeit import default_timer
 from datetime import datetime, timedelta
 from scipy import optimize, interpolate, ndimage
-
-import emg3d  # emg3d for Versions
-
-# Optional modules
-try:
-    import IPython
-except ImportError:
-    IPython = False
-try:
-    import matplotlib
-except ImportError:
-    matplotlib = False
-try:
-    import mkl
-except ImportError:
-    mkl = False
-
-# Get mkl info, if available
-if mkl:
-    mklinfo = mkl.get_version_string()
-else:
-    mklinfo = False
-
 
 __all__ = ['Model', 'Field', 'get_domain', 'get_stretched_h', 'get_hx',
            'get_source_field', 'get_receiver', 'get_h_field', 'TensorMesh',
@@ -1435,26 +1406,20 @@ def data_read(fname, keys=None, path="data"):
 
 
 # OTHER
-class Versions:
+class Versions(scooby.Report):
     r"""Print date, time, and version information.
 
-    Print date, time, and package version information in any environment
-    (Jupyter notebook, IPython console, Python console, QT console), either as
-    html-table (notebook) or as plain text (anywhere).
+    Use scooby to print date, time, and package version information in any
+    environment (Jupyter notebook, IPython console, Python console, QT
+    console), either as html-table (notebook) or as plain text (anywhere).
 
-    Always shown are the OS, number of CPU(s), ``numpy``, ``scipy``,
-    ``emg3d``, ``numba``, ``sys.version``, and time/date.
+    Always shown are the OS, number of CPU(s), ``numpy``, ``scipy``, ``emg3d``,
+    ``numba``, ``sys.version``, and time/date.
 
     Additionally shown are, if they can be imported, ``IPython`` and
     ``matplotlib``. It also shows MKL information, if available.
 
-    All modules provided in ``add_pckg`` are also shown. They have to be
-    imported before ``versions`` is called.
-
-    This script was heavily inspired by:
-
-        - ipynbtools.py from qutip https://github.com/qutip
-        - watermark.py from https://github.com/rasbt/watermark
+    All modules provided in ``add_pckg`` are also shown.
 
 
     Parameters
@@ -1464,8 +1429,14 @@ class Versions:
         imported beforehand).
 
     ncol : int, optional
-        Number of package-columns in html table; only has effect if
-        ``mode='HTML'`` or ``mode='html'``. Defaults to 3.
+        Number of package-columns in html table (no effect in text-version);
+        Defaults to 3.
+
+    text_width : int, optional
+        The text width for non-HTML display modes
+
+    sort : bool, optional
+        Sort the packages when the report is shown
 
 
     Examples
@@ -1479,139 +1450,13 @@ class Versions:
 
     """
 
-    def __init__(self, add_pckg=None, ncol=4):
-        """Initiate and add packages and number of columns to self."""
-        self.add_pckg = self._get_packages(add_pckg)
-        self.ncol = int(ncol)
+    def __init__(self, add_pckg=None, ncol=3, text_width=80, sort=False):
+        """Initiate a scooby.Report instance."""
 
-    def __repr__(self):
-        r"""Plain-text version information."""
+        # Mandatory packages.
+        core = ['numpy', 'scipy', 'numba', 'emg3d']
 
-        # Width for text-version
-        n = 54
-        text = '\n' + n*'-' + '\n'
+        # Optional packages.
+        optional = ['IPython', 'matplotlib']
 
[…]
