+++ .travis.yml
@@ -7,10 +7,10 @@ language: python
 # bookkeeping, so it prints the same version in the Travis info.
 matrix:
   include:
-    - python: 3.7  # => 3.7 with conda-forge; IPython, discretize, no JIT
-      env: PYTHON=3.7 PCKGS="numba IPython" NUMBA_DISABLE_JIT=1 CHAN="conda-forge" INST="discretize"
-    - python: 3.7  # => 3.7 with defaults; mkl
-      env: PYTHON=3.7 PCKGS="mkl mkl-service numba" CHAN="defaults" INST=""
+    - python: 3.7  # => 3.7 no JIT
+      env: PYTHON=3.7 NUMBA_DISABLE_JIT=1
+    - python: 3.7  # => 3.7
+      env: PYTHON=3.7
 
 install:
   - sudo apt-get update
@@ -26,9 +26,9 @@ install:
   - conda info -a
 
   # Install and activate environment, install packages
-  - conda create -q -n test-environment -c $CHAN python=$PYTHON numpy scipy pytest pytest-cov $PCKGS
-  - source activate test-environment
-  - pip install coveralls pytest-flake8 $INST
+  - conda create -q -n tenv python=$PYTHON numpy scipy pytest pytest-cov numba
+  - source activate tenv
+  - pip install coveralls pytest-flake8 scooby
   - python setup.py install
 
 
