### 2019-07-01 #1463 @jedbrown
@prisae Yes, understood that it isn't a feature of emg3d, but a researcher might want to use emg3d in conjunction with other software for that purpose.  We can still ask whether it can do what would be needed for such uses without getting in the way of aspects that are out of scope.  It isn't so much about the particular case of joint inversion, but that its demands are representative of other research applications that may go beyond the scope of the simulation tool and direct uses planned by authors.  Note that such capability is by no means mandatory for publication, but this is a good opportunity for constructive feedback.
