### 2019-07-02 900a086
Make scooby soft dependency
