### 2019-07-03 #1463 @akelbert
Basically, just not my environment. As I said, I'm not a Python person!
So obvious things aren't obvious for me. In trying to update the
packages to

conda install -c prisae emg3d

I completely broke my anaconda so cannot update everything else that I
need to install emg3d.  I know this doesn't belong in this review, but
what do I do when

akelbert$ conda install -c prisae emg3d
Traceback (most recent call last):
   File
"/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages/conda/exceptions.py",
line 1043, in __call__
     return func(*args, **kwargs)
   File
"/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages/conda/cli/main.py",
line 84, in _main
     exit_code = do_call(args, p)
   File
"/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages/conda/cli/conda_argparse.py",
line 80, in do_call
     module = import_module(relative_mod, __name__.rsplit('.', 1)[0])
   File
"/Users/akelbert/Developer/local/miniconda3/lib/python3.6/importlib/__init__.py",
line 126, in import_module
   File "<frozen importlib._bootstrap>", line 978, in _gcd_import
   File "<frozen importlib._bootstrap>", line 961, in _find_and_load
   File "<frozen importlib._bootstrap>", line 950, in
_find_and_load_unlocked
   File "<frozen importlib._bootstrap>", line 655, in _load_unlocked
   File "<frozen importlib._bootstrap_external>", line 678, in exec_module
   File "<frozen importlib._bootstrap>", line 205, in
_call_with_frames_removed
   ...
   File
"/Users/akelbert/Developer/local/miniconda3/lib/python3.6/ctypes/__init__.py",
line 366, in __getitem__
     func = self._FuncPtr((name_or_ordinal, self))
AttributeError: dlsym(0x7f8173d645e0, archive_read_open_filename_w):
symbol not found

Also, this review environment is new to me and I couldn't find an
explanation of how to use it. I don't even spend time on github (use
other types of version control). All the above, conflated with an insane
amount of travel... bad combination.

If I could easily install emg3d I'd be happy to spend some time running
it and providing feedback. Dear authors, I sincerely apologize for
making you wait! Unless @jedbrown is able to help me out, you'd be
better off proceeding without my input for now.
