### 2019-07-03 #1463 @jedbrown
Thanks, @akelbert. I usually recommend pip over anaconda because it's so common for conda users to accidentally mix up which Python is used. Can you try `python3 -m pip install --user emg3d`? This worked well for me. There are tests in the main repository, but the examples might be a better place for you to look at usage. If you have Jupyter (it can also be installed using pip), you can run example notebooks here: https://github.com/empymod/emg3d-examples

These examples depend on the [discretize](https://pypi.org/project/discretize/) package, which can be installed via `python3 -m pip install --user discretize`. I cloned the examples repository, opened the [minimal working example](https://github.com/empymod/emg3d-examples/blob/master/0a_Minimum_working_example.ipynb), and was able to reproduce successfully, for example.

(@prisae I did encounter `Javascript Error: IPython is not defined` on cells 5 and 8 when using Jupyter Lab; it works fine with Jupyter Notebook. Is there a missing import/magic or does that widget not work with Lab?)
