### 2019-07-03 #1463 @akelbert
@jedbrown, thanks for the response! Any ideas?

akelbert$ python3 -m pip install --user emg3d
Collecting emg3d
   Could not fetch URL https://pypi.python.org/simple/emg3d/: There was
a problem confirming the ssl certificate: [SSL:
CERTIFICATE_VERIFY_FAILED] certificate verify failed (_ssl.c:749) - skipping
   Could not find a version that satisfies the requirement emg3d (from
versions: )
No matching distribution found for emg3d
