### 2019-07-03 #1463 @jedbrown
For our reference, can you paste the output from `python3 -m pip --version`?
I think [this answer](https://stackoverflow.com/a/29751768/33208) will resolve the cert issue.
