### 2019-07-03 #1463 @akelbert
akelbert$ python3 -m pip --version
pip 9.0.1 from
/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages
(python 3.6)

akelbert$ pip install --trusted-host pypi.org --trusted-host
files.pythonhosted.org emg3d
Collecting emg3d
   Could not fetch URL https://pypi.python.org/simple/emg3d/: There was
a problem confirming the ssl certificate: [SSL:
CERTIFICATE_VERIFY_FAILED] certificate verify failed (_ssl.c:749) - skipping
   Could not find a version that satisfies the requirement emg3d (from
versions: )
No matching distribution found for emg3d

akelbert$ pip install --trusted-host pypi.org --trusted-host
files.pythonhosted.org pip setuptools
Requirement already satisfied: pip in
/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages
Requirement already satisfied: setuptools in
/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages

Somehow I've never had luck with python :(
