### 2019-07-03 #1463 @jedbrown
Hmm, can you try this?
`pip install --upgrade --user --trusted-host pypi.org --trusted-host files.pythonhosted.org pip setuptools`
