### 2019-07-03 #1463 @akelbert
akelbert$ pip install --upgrade --user --trusted-host pypi.org
--trusted-host files.pythonhosted.org pip setuptools
Could not fetch URL https://pypi.python.org/simple/pip/: There was a
problem confirming the ssl certificate: [SSL: CERTIFICATE_VERIFY_FAILED]
certificate verify failed (_ssl.c:749) - skipping
Requirement already up-to-date: pip in
/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages
Could not fetch URL https://pypi.python.org/simple/setuptools/: There
was a problem confirming the ssl certificate: [SSL:
CERTIFICATE_VERIFY_FAILED] certificate verify failed (_ssl.c:749) - skipping
Requirement already up-to-date: setuptools in
/Users/akelbert/Developer/local/miniconda3/lib/python3.6/site-packages

:)
