### 2019-07-03 #1463 @prisae
Yes, emg3d requires python 3.7 at least.

As it is a new library I decided to use the newest python to leverage some of the new tools in python.

I'll be in the office in an hour or so and will reply to the questions raised in the last few comments.
