### 2019-07-03 #1463 @prisae
I am not sure about the SSL stuff, as this is more OS/network related, but I can do some digging.

However, could you try this:
```
wget https://repo.continuum.io/archive/Anaconda3-2019.03-MacOSX-x86_64.sh
bash Anaconda3-2019.03-MacOSX-x86_64.sh -b -p $HOME/testanaconda
export PATH="$HOME/testanaconda/bin:$PATH"
conda install -c prisae emg3d
```
I hope it works on a Mac, I just tried it on Linux and adjusted the filename.

This will install the most recent anaconda (Python3) in your home in `testanaconda`, and also install `emg3d`.
