### 2019-07-03 #1463 @prisae
If it does work, here would be a minimal example:

Start IPython (or pure Python if you want)
```bash
ipython
```

And then, in the IPython shell:
```py
import emg3d
import numpy as np
grid = emg3d.utils.TensorMesh([np.ones(8)*10, np.ones(8)*10, np.ones(8)*10], x0=[-40, -40, -40])
freq = 10.0
model = emg3d.utils.Model(grid, res_x=1.5, freq=freq)
sfield = emg3d.utils.get_source_field(grid, src=[0, 0, 0, 0, 0], freq=freq)
efield = emg3d.solver.solver(grid, model, sfield, verb=3)
```

Very basic example, but then we would know it works.

The output should be along the lines of:
```
:: emg3d START :: 08:25:26 ::

   MG-cycle       : 'F'                 sslsolver : False
   semicoarsening : False [0]           tol       : 1e-06
   linerelaxation : False [0]           maxit     : 50
   nu_{i,1,c,2}   : 0, 2, 1, 2          verb      : 3
   Original grid  :   8 x   8 x   8     => 512 cells
   Coarsest grid  :   2 x   2 x   2     => 8 cells
   Coarsest level :   2 ;   2 ;   2

   [hh:mm:ss]     error                 l2:[last/init, last/prev] l s

       h_
      2h_ \    /
      4h_  \/\/

   [08:25:26] 1.420e-06 after  1 F-cycles; [2.544e-02, 2.544e-02] 0 0
   [08:25:26] 1.111e-07 after  2 F-cycles; [1.990e-03, 7.821e-02] 0 0
   [08:25:26] 1.026e-08 after  3 F-cycles; [1.838e-04, 9.239e-02] 0 0
   [08:25:26] 1.006e-09 after  4 F-cycles; [1.803e-05, 9.806e-02] 0 0
   [08:25:26] 1.014e-10 after  5 F-cycles; [1.816e-06, 1.007e-01] 0 0
   [08:25:26] 1.036e-11 after  6 F-cycles; [1.855e-07, 1.021e-01] 0 0

   > CONVERGED
   > MG cycles      : 6
   > Final l2-norm  : 1.036e-11

:: emg3d END   :: 08:25:26 :: runtime = 0:00:00
```
