### 2019-07-03 #1463 @akelbert
Thank you, sounds great! I'll try your suggestions tomorrow, it's past
midnight here. One of the challenges is that as a government employee in
the US, I have no admin permissions on my own machine. Makes things
difficult.

If I can make this work tomorrow, I'll run the examples and respond
immediately. If it doesn't work... well, then it doesn't. I'll let you
know either way.
