### 2019-07-03 #1463 @prisae
@jedbrown regarding the issue with JupyterLab:

This is probably one of the reasons why I am still sticking around with notebooks, and not with the lab. Interactive figures in JupyterLab are not that straight forward.

You could either:
- Change `%matplotlib notebook` to `%matplotlib inline` in the 2nd cell. Then the figures are shown, but they won't be interactive.
- Follow https://github.com/matplotlib/jupyter-matplotlib to install all the required dependencies which will enable interactive matplotlib figures in JupyterLab (with `%matplotlib widget`).

I tried both version just now, and they seem to work.
