### 2019-07-03 #1463 @prisae
@akelbert That is fine. You don't need admin permission to install/run stuff in python, that is not an issue. However, firewall might be an issue...
