### 2019-07-03 d221fad
Make scooby soft dependency
