### 2019-07-03 #1463 @prisae
@akelbert : If there is an issue with installing it you could use Binder instead just to check a few examples. Click on the batch for the example notebook `0a_Minimum_working_example` [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/empymod/emg3d-examples/master?filepath=0a_Minimum_working_example.ipynb)

The only thing you have to add to run it on Binder is to add the following two lines:
```
!pip install scipy numpy numba emg3d matplotlib
!pip install discretize
```
to a new cell at the beginning. For this click on the plus-symbol just next to the save-symbol. Running this cell will then install everything (it will take some time...).

When running the actual calculation it will take much longer the first time, as it has to compile some parts. Consecutive runs (in the same notebook instance) will be faster.
