### 2019-07-03 #1463 @prisae
In this way you could run any notebook; this batch opens you the whole `emg3d-examples`-directory on Binder, so you can select a notebook: [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/empymod/emg3d-examples/master)

Depending on the examples you might have to adjust the `!pip install`-commands given above for additional dependencies.
