### 2019-07-03 05d13b2
Jit np.linalg.norm to njited.l2norm
