### 2019-07-03 8e82ce6
Adjust test re previous commit
