### 2019-07-03 12d46f6
Try to fix coveralls issue
