+++ .travis.yml
@@ -9,6 +9,8 @@ matrix:
   include:
     - python: 3.7  # => 3.7 no JIT; scooby
       env: PYTHON=3.7 NUMBA_DISABLE_JIT=1 INST="scooby"
+    - python: 3.7  # => Will become 3.8; currently just to fix coveralls
+      env: PYTHON=3.7 NUMBA_DISABLE_JIT=1 INST="scooby"
     - python: 3.7  # => 3.7
       env: PYTHON=3.7 INST=""
 
