### 2019-07-04 #1463 @akelbert
Sorry Dieter, nothing worked for me yet, and it's a holiday season here
in the US - I'm leaving on vacation in the morning. Will try the last
few suggestions on Monday. Thanks for being so patient with me. This is
also a use case of sorts: an EM modeler who's never done python...
