### 2019-07-04 c861d81
Improve error log; also speed
This coincidently should also improve speed, as we calculated the
reference error l2_refe in each recursive cycle, where we only need it
once.
