+++ emg3d/solver.py
@@ -312,8 +312,13 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
             linerelaxation=linerelaxation, vnC=grid.vnC, verb=verb, **kwargs
     )
 
-    # Start logging and print all parameters.
+    # Start logging
     var.cprint(f"\n:: emg3d START :: {var.time.now} ::\n", 1)
+
+    # Calculate reference norm for tolerance.
+    var.l2_refe = njitted.l2norm(sfield)
+
+    # Print all parameters.
     var.cprint(var, 1)
 
     # Get efield
@@ -333,7 +338,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
 
         # If efield is provided, check if it is already sufficiently good.
         var.l2 = residual(grid, model, sfield, efield, True)
-        if var.l2 < var.tol*njitted.l2norm(sfield):
+        if var.l2 < var.tol*var.l2_refe:
 
             # Switch-off both sslsolver and multigrid.
             var.sslsolver = None
@@ -343,14 +348,14 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
             info = f"   > NOTHING DONE (provided efield already good enough)\n"
 
     # Print header for iteration log.
-    header = f"   [hh:mm:ss]     {'error':<15}"
+    header = f"   [hh:mm:ss]  {'rel. error':<18}"
     if var.sslsolver:
         header += f"{'solver':<20} "
         if var.cycle:
             header += f"{'MG':<11} l s"
         var.cprint(header+"\n", 2)
     elif var.cycle:
-        var.cprint(header+f"{'l2:[last/init, last/prev]':>32} l s\n", 2)
+        var.cprint(header+f"{'l2:[absolut, last/prev]':>32} l s\n", 2)
 
     # Solve the system with...
     if var.sslsolver:  # ... sslsolver.
@@ -365,7 +370,8 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
             info += f"   > MG prec. steps : {var.it}\n"
     elif var.cycle:    # multigrid-specific info.
         info = f"   > MG cycles      : {var.it}\n"
-    info += f"   > Final l2-norm  : {var.l2:.3e}\n\n"  # Final error.
+    info += f"   > Final error    : {var.l2:.3e}"      # Final error.
+    info += f" ({var.l2/var.l2_refe:.3e} rel.)\n\n"    # Relative.
     info += f":: emg3d END   :: {var.time.now} :: "    # END and time.
     info += f"runtime = {var.time.runtime}\n"          # Total runtime.
     var.cprint(info, 1)
@@ -437,11 +443,8 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
         cycmax = new_cycmax
     cyc = 0  # Initiate cycle count.
 
-    # Define various l2-norms.
-    l2_refe = njitted.l2norm(sfield)  # Reference norm for tolerance.
-    l2_last = residual(grid, model, sfield, efield, True)  # Current residual.
-    l2_init = l2_last
-    l2_prev = l2_last
+    # Calculate current error (l2-norms).
+    l2_last = residual(grid, model, sfield, efield, True)
 
     # Keep track on the levels during the first cycle, for QC.
     if var._first_cycle:
@@ -539,7 +542,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
             l2_last = residual(grid, model, sfield, efield, True)
 
             # Print end-of-cycle info.
-            _print_cycle_info(var, l2_last, l2_prev, l2_init)
[…]
