### 2019-07-04 435bd36
Adjust tests
