### 2019-07-04 8ec69f1
Simplify
