+++ tests/test_solver.py
@@ -40,7 +40,7 @@ def test_solver_homogeneous(capsys):
     assert ' emg3d START ::' in out
     assert ' [hh:mm:ss] ' in out
     assert ' MG cycles ' in out
-    assert ' Final error ' in out
+    assert ' Final rel. error ' in out
     assert ' emg3d END   :: ' in out
 
     # Check all fields (ex, ey, and ez)
@@ -67,7 +67,7 @@ def test_solver_homogeneous(capsys):
     assert ' CONVERGED' in out
     assert ' Solver steps ' in out
     assert ' MG prec. steps ' in out
-    assert ' Final error ' in out
+    assert ' Final rel. error ' in out
     assert ' emg3d END   :: ' in out
 
     # Check all fields (ex, ey, and ez)
@@ -80,7 +80,7 @@ def test_solver_homogeneous(capsys):
     assert ' [hh:mm:ss] ' in out
     assert ' CONVERGED' in out
     assert ' MG cycles ' in out
-    assert ' Final error ' in out
+    assert ' Final rel. error ' in out
     assert ' emg3d END   :: ' in out
 
     # Max it
@@ -340,12 +340,12 @@ def test_krylov(capsys):
     efield = utils.Field(grid)  # Initiate e-field.
 
     # Get var-instance
-    l2_refe = njitted.l2norm(sfield)
     var = solver.MGParameters(
             cycle=None, sslsolver=True, semicoarsening=False,
-            linerelaxation=False, vnC=grid.vnC, verb=3, l2_refe=l2_refe,
+            linerelaxation=False, vnC=grid.vnC, verb=3,
             maxit=-1,  # Set stupid input to make bicgstab fail.
     )
+    var.l2_refe = njitted.l2norm(sfield)
 
     # Call krylov and ensure it fails properly.
     solver.krylov(grid, model, sfield, efield, var)
@@ -358,67 +358,62 @@ def test_mgparameters():
 
     # 1. semicoarsening
     var = solver.MGParameters(cycle='F', sslsolver=False, semicoarsening=True,
-                              linerelaxation=False, vnC=vnC, verb=1, l2_refe=1)
+                              linerelaxation=False, vnC=vnC, verb=1)
     assert 'semicoarsening : True [1 2 3]' in var.__repr__()
-    assert 'ref. error : 1' in var.__repr__()
     var = solver.MGParameters(cycle='V', sslsolver=False, semicoarsening=1213,
-                              linerelaxation=False, vnC=vnC, verb=1, l2_refe=1)
+                              linerelaxation=False, vnC=vnC, verb=1)
     assert 'semicoarsening : True [1 2 1 3]' in var.__repr__()
     var = solver.MGParameters(cycle='F', sslsolver=False, semicoarsening=2,
-                              linerelaxation=False, vnC=vnC, verb=1, l2_refe=1)
+                              linerelaxation=False, vnC=vnC, verb=1)
     assert 'semicoarsening : True [2]' in var.__repr__()
     with pytest.raises(ValueError):
         solver.MGParameters(cycle='F', sslsolver=False, semicoarsening=5,
-                            linerelaxation=False, vnC=vnC, verb=1, l2_refe=1)
+                            linerelaxation=False, vnC=vnC, verb=1)
 
     # 2. linerelaxation
     var = solver.MGParameters(cycle='F', sslsolver=False, semicoarsening=False,
-                              linerelaxation=True, vnC=vnC, verb=1, l2_refe=1)
+                              linerelaxation=True, vnC=vnC, verb=1)
     assert 'linerelaxation : True [4 5 6]' in var.__repr__()
     var = solver.MGParameters(cycle='F', sslsolver=False, semicoarsening=False,
-                              linerelaxation=1247, vnC=vnC, verb=1, l2_refe=1)
[…]
