+++ emg3d/solver.py
@@ -167,7 +167,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         - 1: Print warnings.
         - 2: Print runtime and information about the method.
         - 3: Print additional information for each MG-cycle.
-        - 4: Print everything (slower due to additional l2-norm calculations).
+        - 4: Print everything (slower due to additional error calculations).
 
     **kwargs : Optional solver options:
 
@@ -227,8 +227,8 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         initial efield was provided.
 
     info_dict : dict
-        Dictionary with runtime info (final norm and number of iterations of MG
-        and the sslsolver); only if ``return_info=True``.
+        Dictionary with runtime info (errors and number of iterations); only if
+        ``return_info=True``.
 
 
     Examples
@@ -279,7 +279,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
        Coarsest grid  :   3 x   2 x   2     => 12 cells
        Coarsest level :   4 ;   4 ;   4
     .
-       [hh:mm:ss]     error                 l2:[last/init, last/prev] l s
+       [hh:mm:ss]  rel. error              error:[absolut, last/prev] l s
     .
            h_
           2h_ \                  /
@@ -287,17 +287,17 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
           8h_   \    /\  /  \  /
          16h_    \/\/  \/    \/
     .
-       [15:24:40] 1.464e-06 after  1 F-cycles; [2.623e-02, 2.623e-02] 0 0
-       [15:24:40] 1.258e-07 after  2 F-cycles; [2.253e-03, 8.589e-02] 0 0
-       [15:24:41] 1.704e-08 after  3 F-cycles; [3.051e-04, 1.354e-01] 0 0
-       [15:24:41] 3.071e-09 after  4 F-cycles; [5.500e-05, 1.803e-01] 0 0
-       [15:24:41] 6.531e-10 after  5 F-cycles; [1.170e-05, 2.127e-01] 0 0
-       [15:24:42] 1.532e-10 after  6 F-cycles; [2.745e-06, 2.346e-01] 0 0
-       [15:24:42] 3.837e-11 after  7 F-cycles; [6.873e-07, 2.504e-01] 0 0
+       [15:24:40] 2.623e-02 after  1 F-cycles; [1.464e-06, 2.623e-02] 0 0
+       [15:24:40] 2.253e-03 after  2 F-cycles; [1.258e-07, 8.589e-02] 0 0
+       [15:24:41] 3.051e-04 after  3 F-cycles; [1.704e-08, 1.354e-01] 0 0
+       [15:24:41] 5.500e-05 after  4 F-cycles; [3.071e-09, 1.803e-01] 0 0
+       [15:24:41] 1.170e-05 after  5 F-cycles; [6.531e-10, 2.127e-01] 0 0
+       [15:24:42] 2.745e-06 after  6 F-cycles; [1.532e-10, 2.346e-01] 0 0
+       [15:24:42] 6.873e-07 after  7 F-cycles; [3.837e-11, 2.504e-01] 0 0
     .
        > CONVERGED
-       > MG cycles      : 7
-       > Final l2-norm  : 3.837e-11
+       > MG cycles        : 7
+       > Final rel. error : 6.873e-07
     .
     :: emg3d END :: 15:24:42 :: runtime = 0:00:02.177778
 
@@ -316,7 +316,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
[…]
