### 2019-07-05 48240d5
minimize diff
