### 2019-07-05 3dfbf2a
Improve error log
This coincidently should also improve speed a tiny bit, as we calculated the reference error `l2_refe` in each recursive cycle, where we only need it once.
