### 2019-07-08 8cbe8de
Adjust changelog
