+++ emg3d/utils.py
@@ -40,8 +40,8 @@ except ImportError:
                   "\n             Install it via `pip install scooby`.\n")
 
 __all__ = ['Model', 'Field', 'get_domain', 'get_stretched_h', 'get_hx',
-           'get_source_field', 'get_receiver', 'get_h_field', 'TensorMesh',
-           'Time', 'data_write', 'data_read', 'Report']
+           'get_source_field', 'get_receiver', 'get_h_field', 'grid2grid',
+           'TensorMesh', 'Time', 'data_write', 'data_read', 'Report']
 
 
 # CONSTANTS
@@ -1152,12 +1152,6 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
 
     If ``rec_loc`` is outside of ``grid``, the returned field is zero.
 
-    This is a modified version of :func:`scipy.interpolate.interpn`, using
-    :class:`scipy.interpolate.RegularGridInterpolator` if ``method='linear'``
-    and a custom-wrapped version of :func:`scipy.ndimage.map_coordinates` if
-    ``method='cubic'``. If speed is important then choose 'linear', as it can
-    be significantly faster.
-
 
     Parameters
     ----------
@@ -1212,50 +1206,53 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
         # Add to points
         points += pts
 
-        # We need at least 3 points in each direction for cubic spline.
-        # This should never be an issue for a realistic 3D model.
-        if pts[0].size < 4:
-            method = 'linear'
+    return _interp3d(points, fieldf, rec_loc, method, 0.0, 'constant')
 
-    # Interpolation.
-    if method == "linear":
-        ifn = interpolate.RegularGridInterpolator(
-                points=points, values=fieldf, method="linear",
-                bounds_error=False, fill_value=0.0)
 
-        return ifn(xi=rec_loc)
+def grid2grid(grid_in, values_in, grid_out, method='cubic'):
+    """Interpolate values from one to another grid.
 
-    else:
+    Return ``values_in`` corresponding to ``grid_in`` on the points in
+    ``grid_out``.
 
-        # Replicate the same expansion of xi as used in
-        # RegularGridInterpolator, so the input xi can be quite flexible.
-        xi = interpolate.interpnd._ndim_coords_from_arrays(rec_loc, ndim=3)
-        xi_shape = xi.shape
-        xi = xi.reshape(-1, 3)
+    Points on ``grid_out`` which are outside of ``grid_in`` are extrapolated
+    (if ``method='linear'``) or the nearest value is taken (if
+    ``method='cubic'``).
 
-        # map_coordinates uses the indices of the input data (fieldf in this
-        # case) as coordinates. We have therefore to transform our desired
-        # output coordinates to this artificial coordinate system too.
-        params = {'kind': 'cubic',
-                  'bounds_error': False,
-                  'fill_value': 'extrapolate'}
-        x = interpolate.interp1d(
-                points[0], np.arange(len(points[0])), **params)(xi[:, 0])
-        y = interpolate.interp1d(
-                points[1], np.arange(len(points[1])), **params)(xi[:, 1])
-        z = interpolate.interp1d(
-                points[2], np.arange(len(points[2])), **params)(xi[:, 2])
-        coords = np.vstack([x, y, z])
 
-        # map_coordinates only works for real data; split it up if complex.
-        if 'complex' in fieldf.dtype.name:
-            real = ndimage.map_coordinates(fieldf.real, coords, order=3)
-            imag = ndimage.map_coordinates(fieldf.imag, coords, order=3)
-            result = real + 1j*imag
-        else:
-            result = ndimage.map_coordinates(fieldf, coords, order=3)
+    Parameters
+    ----------
+    grid_in, grid_out : TensorMesh
[…]
