+++ emg3d/utils.py
@@ -1256,9 +1256,9 @@ def grid2grid(grid_in, values_in, grid_out, method='cubic'):
         out_points = grid_out.gridCC
     else:
         xx, yy, zz = np.broadcast_arrays(
-                grid_out.vectorCCx[:, None, None],
-                grid_out.vectorCCy[None, :, None],
-                grid_out.vectorCCz[None, None, :])
+                grid_out.vectorCCx,
+                grid_out.vectorCCy[:, None],
+                grid_out.vectorCCz[:, None, None])
         out_points = np.r_[xx.ravel('F'), yy.ravel('F'), zz.ravel('F')]
         out_points = out_points.reshape(-1, 3, order='F')
 
@@ -1564,7 +1564,6 @@ def _interp3d(points, values, new_points, method, fill_value=0.0,
                 points[2], np.arange(len(points[2])), **params1d)(xi[:, 2])
         coords = np.vstack([x, y, z])
 
-        print(mode)
         # map_coordinates only works for real data; split it up if complex.
         params3d = {'order': 3, 'mode': mode, 'cval': 0.0}
         if 'complex' in values.dtype.name:
