### 2019-07-08 #1463 @akelbert
Your basic example worked for me after I reinstalled Anaconda like you suggested. The Binder also worked. I'm leaving for a conference in the morning and have a bunch of other things to work through, but at least I'm now equipped with the tools to run the code and play with the settings. Will do so in between everything else that I'm playing with this week (giving a talk, chairing a session, etc). Thank you so much for your patience!!!
