### 2019-07-09 6ead79a
New fct grid2grid
New function `utils.grid2grid` to move between different grids with either linear or cubic spline interpolation.
