+++ emg3d/utils.py
@@ -40,8 +40,8 @@ except ImportError:
                   "\n             Install it via `pip install scooby`.\n")
 
 __all__ = ['Model', 'Field', 'get_domain', 'get_stretched_h', 'get_hx',
-           'get_source_field', 'get_receiver', 'get_h_field', 'TensorMesh',
-           'Time', 'data_write', 'data_read', 'Report']
+           'get_source_field', 'get_receiver', 'get_h_field', 'grid2grid',
+           'TensorMesh', 'Time', 'data_write', 'data_read', 'Report']
 
 
 # CONSTANTS
@@ -1152,12 +1152,6 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
 
     If ``rec_loc`` is outside of ``grid``, the returned field is zero.
 
-    This is a modified version of :func:`scipy.interpolate.interpn`, using
-    :class:`scipy.interpolate.RegularGridInterpolator` if ``method='linear'``
-    and a custom-wrapped version of :func:`scipy.ndimage.map_coordinates` if
-    ``method='cubic'``. If speed is important then choose 'linear', as it can
-    be significantly faster.
-
 
     Parameters
     ----------
@@ -1212,50 +1206,64 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
         # Add to points
         points += pts
 
-        # We need at least 3 points in each direction for cubic spline.
-        # This should never be an issue for a realistic 3D model.
-        if pts[0].size < 4:
-            method = 'linear'
+    return _interp3d(points, fieldf, rec_loc, method, 0.0, 'constant')
 
-    # Interpolation.
-    if method == "linear":
-        ifn = interpolate.RegularGridInterpolator(
-                points=points, values=fieldf, method="linear",
-                bounds_error=False, fill_value=0.0)
 
-        return ifn(xi=rec_loc)
+def grid2grid(grid_in, values_in, grid_out, method='linear'):
+    """Interpolate values from one to another grid.
 
+    Return ``values_in`` corresponding to ``grid_in`` on the points in
+    ``grid_out``.
+
+    Points on ``grid_out`` which are outside of ``grid_in`` are extrapolated
+    (if ``method='linear'``) or the nearest value is taken (if
+    ``method='cubic'``).
+
+
+    Parameters
+    ----------
+    grid_in, grid_out : TensorMesh
+        Input and output model grids; ``TensorMesh``-instances.
+
+    values_in : ndarray
+        Values corresponding to ``grid_in``.
+
+    method : str, optional
+        The method of interpolation to perform, 'linear' or 'cubic'.
+        Default is 'cubic' (forced to 'linear' if there are less than 3 points
+        in any direction).
+
+
+    Returns
+    -------
+    values_out : ndarray
+        Values corresponding to ``grid_out``.
+
+    """
+    # Ensure values has the dimensions of grid_in.
+    if not np.all(grid_in.vnC == values_in.shape):
+        print("* ERROR   :: ``values_in`` must have same shape as "
+              "``grid_in``.")
+        print(f"             Shape of ``grid_in``   : {grid_in.vnC}.")
+        print(f"             Shape of ``values_in`` : {values_in.shape}.")
+        raise ValueError("grid_in or value_in error")
+
+    # Get the vectors corresponding to input data.
+    points = (grid_in.vectorCCx, grid_in.vectorCCy, grid_in.vectorCCz)
+
+    # Format the output points.
+    if hasattr(grid_out, 'gridCC'):
+        out_points = grid_out.gridCC
     else:
+        xx, yy, zz = np.broadcast_arrays(
+                grid_out.vectorCCx[:, None, None],
[…]
