+++ emg3d/__init__.py
@@ -29,4 +29,4 @@ from .utils import Report
 
 __all__ = ['solver', 'utils', 'Report']
 
-__version__ = '0.7.1'
+__version__ = '0.7.1dev0'
