### 2019-07-12 f776b53
working version; not jet jitted
