+++ emg3d/utils.py
@@ -1209,7 +1209,7 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
     return _interp3d(points, fieldf, rec_loc, method, 0.0, 'constant')
 
 
-def grid2grid(grid_in, values_in, grid_out, method='linear'):
+def grid2grid(grid_in, values_in, grid_out, method='volume'):
     """Interpolate values from one to another grid.
 
     Return ``values_in`` corresponding to ``grid_in`` on the points in
@@ -1217,7 +1217,7 @@ def grid2grid(grid_in, values_in, grid_out, method='linear'):
 
     Points on ``grid_out`` which are outside of ``grid_in`` are extrapolated
     (if ``method='linear'``) or the nearest value is taken (if
-    ``method='cubic'``).
+    ``method='volume'`` or ``method='cubic'``).
 
 
     Parameters
@@ -1228,10 +1228,12 @@ def grid2grid(grid_in, values_in, grid_out, method='linear'):
     values_in : ndarray
         Values corresponding to ``grid_in``.
 
-    method : str, optional
-        The method of interpolation to perform, 'linear' or 'cubic'.
-        Default is 'cubic' (forced to 'linear' if there are less than 3 points
-        in any direction).
+    method : {<'volume'>, 'linear', 'cubic'}, optional
+        The method of interpolation to perform.
+        The volume averaging method ensures that the total sum of the property
+        stays constant. Default is 'volume'. The method 'cubic' requires at
+        least three points in any direction, otherwise it will fall back to
+        'linear'.
 
 
     Returns
@@ -1248,22 +1250,30 @@ def grid2grid(grid_in, values_in, grid_out, method='linear'):
         print(f"             Shape of ``values_in`` : {values_in.shape}.")
         raise ValueError("grid_in or value_in error")
 
-    # Get the vectors corresponding to input data.
-    points = (grid_in.vectorCCx, grid_in.vectorCCy, grid_in.vectorCCz)
+    if method == 'volume':
+        points_in = (grid_in.vectorNx, grid_in.vectorNy, grid_in.vectorNz)
+        points_out = (grid_out.vectorNx, grid_out.vectorNy, grid_out.vectorNz)
+        values_out = _volume_average(*points_in, values_in, *points_out)
 
-    # Format the output points.
-    if hasattr(grid_out, 'gridCC'):
-        out_points = grid_out.gridCC
     else:
-        xx, yy, zz = np.broadcast_arrays(
-                grid_out.vectorCCx[:, None, None],
-                grid_out.vectorCCy[:, None],
-                grid_out.vectorCCz)
-        out_points = np.r_[xx.ravel('F'), yy.ravel('F'), zz.ravel('F')]
-        out_points = out_points.reshape(-1, 3, order='F')
+        # Get the vectors corresponding to input data.
+        points = (grid_in.vectorCCx, grid_in.vectorCCy, grid_in.vectorCCz)
 
-    return _interp3d(
-            points, values_in, out_points, method, None, 'nearest')
+        # Format the output points.
+        if hasattr(grid_out, 'gridCC'):
+            out_points = grid_out.gridCC
+        else:
+            xx, yy, zz = np.broadcast_arrays(
+                    grid_out.vectorCCx[:, None, None],
+                    grid_out.vectorCCy[:, None],
+                    grid_out.vectorCCz)
+            out_points = np.r_[xx.ravel('F'), yy.ravel('F'), zz.ravel('F')]
+            out_points = out_points.reshape(-1, 3, order='F')
+
[…]
