+++ emg3d/utils.py
@@ -30,6 +30,8 @@ from timeit import default_timer
 from datetime import datetime, timedelta
 from scipy import optimize, interpolate, ndimage
 
+from . import njitted
+
 # scooby is a soft dependency for emg3d
 try:
     from scooby import Report as ScoobyReport
@@ -1210,14 +1212,11 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
 
 
 def grid2grid(grid_in, values_in, grid_out, method='volume'):
-    """Interpolate values from one to another grid.
+    """Interpolate ``values_in`` located on ``grid_in`` to ``grid_out``.
 
-    Return ``values_in`` corresponding to ``grid_in`` on the points in
-    ``grid_out``.
-
-    Points on ``grid_out`` which are outside of ``grid_in`` are extrapolated
-    (if ``method='linear'``) or the nearest value is taken (if
-    ``method='volume'`` or ``method='cubic'``).
+    Points on ``grid_out`` which are outside of ``grid_in`` are filled by the
+    nearest value (if ``method='volume'`` or ``method='cubic'``) or by
+    extrapolation (if ``method='linear'``).
 
 
     Parameters
@@ -1242,6 +1241,7 @@ def grid2grid(grid_in, values_in, grid_out, method='volume'):
         Values corresponding to ``grid_out``.
 
     """
+
     # Ensure values has the dimensions of grid_in.
     if not np.all(grid_in.vnC == values_in.shape):
         print("* ERROR   :: ``values_in`` must have same shape as "
@@ -1253,7 +1253,10 @@ def grid2grid(grid_in, values_in, grid_out, method='volume'):
     if method == 'volume':
         points_in = (grid_in.vectorNx, grid_in.vectorNy, grid_in.vectorNz)
         points_out = (grid_out.vectorNx, grid_out.vectorNy, grid_out.vectorNz)
-        values_out = _volume_average(*points_in, values_in, *points_out)
+        values_out = np.zeros(grid_out.vnC, dtype=values_in.dtype)
+
+        # Get values from `volume_average`.
+        njitted.volume_average(*points_in, values_in, *points_out, values_out)
 
     else:
         # Get the vectors corresponding to input data.
@@ -1270,6 +1273,7 @@ def grid2grid(grid_in, values_in, grid_out, method='volume'):
             out_points = np.r_[xx.ravel('F'), yy.ravel('F'), zz.ravel('F')]
             out_points = out_points.reshape(-1, 3, order='F')
 
+        # Get values from `_interp3d`.
         values_out = _interp3d(
                 points, values_in, out_points, method, None, 'nearest')
 
@@ -1582,81 +1586,3 @@ def _interp3d(points, values, new_points, method, fill_value, mode):
             result = ndimage.map_coordinates(values, coords, **params3d)
 
         return result.reshape(xi_shape[:-1])
-
-
-def _volume_average(edges_x_in, edges_y_in, edges_z_in, values_in, edges_x_out,
-                    edges_y_out, edges_z_out):
-    """Interpolate values_in on grid_in to grid_out."""
-
-    # Get cell indices.
-    # First and last edges ignored => first and last cells extend to +/- infty.
-    ix_l = np.searchsorted(edges_x_in[1:-1], edges_x_out, 'left')
-    ix_r = np.searchsorted(edges_x_in[1:-1], edges_x_out, 'right')
-    iy_l = np.searchsorted(edges_y_in[1:-1], edges_y_out, 'left')
-    iy_r = np.searchsorted(edges_y_in[1:-1], edges_y_out, 'right')
[…]
