+++ emg3d/njitted.py
@@ -1999,54 +1999,56 @@ def restrict_weights(vectorN, vectorCC, h, cvectorN, cvectorCC, ch):
 
 # Volume averaging
 @nb.njit(**_numba_setting)
-def volume_average(edges_x_in, edges_y_in, edges_z_in, values_in, edges_x_out,
-                   edges_y_out, edges_z_out, values_out):
+def volume_average(edges_x, edges_y, edges_z, values, new_edges_x, new_edges_y,
+                   new_edges_z, new_values):
     """Interpolation using the volume averaging technique.
 
-    The result is added to values_out.
+    The result is added to new_values.
 
     Parameters
     ----------
-    edges_[x, y, z]_[in, out] : ndarray
-        The edges in x-, y-, and z-directions for the input and the output
-        grids, respectively.
+    edges_[x, y, z] : ndarray
+        The edges in x-, y-, and z-directions for the original grid.
 
-    values_in : ndarray
-        Values corresponding to ``grid_in``.
+    values : ndarray
+        Values corresponding to ``grid``.
 
-    values_out : ndarray
-        Array where values corresponding to ``grid_out`` will be added.
+    new_edges_[x, y, z] : ndarray
+        The edges in x-, y-, and z-directions for the new grid.
+
+    new_values : ndarray
+        Array where values corresponding to ``new_grid`` will be added.
 
     """
 
     # Get cell indices.
     # First and last edges ignored => first and last cells extend to +/- infty.
-    ix_l = np.searchsorted(edges_x_in[1:-1], edges_x_out, 'left')
-    ix_r = np.searchsorted(edges_x_in[1:-1], edges_x_out, 'right')
-    iy_l = np.searchsorted(edges_y_in[1:-1], edges_y_out, 'left')
-    iy_r = np.searchsorted(edges_y_in[1:-1], edges_y_out, 'right')
-    iz_l = np.searchsorted(edges_z_in[1:-1], edges_z_out, 'left')
-    iz_r = np.searchsorted(edges_z_in[1:-1], edges_z_out, 'right')
+    ix_l = np.searchsorted(edges_x[1:-1], new_edges_x, 'left')
+    ix_r = np.searchsorted(edges_x[1:-1], new_edges_x, 'right')
+    iy_l = np.searchsorted(edges_y[1:-1], new_edges_y, 'left')
+    iy_r = np.searchsorted(edges_y[1:-1], new_edges_y, 'right')
+    iz_l = np.searchsorted(edges_z[1:-1], new_edges_z, 'left')
+    iz_r = np.searchsorted(edges_z[1:-1], new_edges_z, 'right')
 
     # Get number of cells.
-    ncx = len(edges_x_out)-1
-    ncy = len(edges_y_out)-1
-    ncz = len(edges_z_out)-1
+    ncx = len(new_edges_x)-1
+    ncy = len(new_edges_y)-1
+    ncz = len(new_edges_z)-1
 
     # Working arrays for edges; ensure they are big enough.
-    x_edges = np.zeros(max(ncx, len(edges_x_in))+3)
-    y_edges = np.zeros(max(ncy, len(edges_y_in))+3)
-    z_edges = np.zeros(max(ncz, len(edges_z_in))+3)
+    x_edges = np.zeros(max(ncx, len(edges_x))+3)
+    y_edges = np.zeros(max(ncy, len(edges_y))+3)
+    z_edges = np.zeros(max(ncz, len(edges_z))+3)
 
-    # Loop over grid_out cells.
+    # Loop over new_grid cells.
     for iz in range(ncz):
-        hz = np.diff(edges_z_out[iz:iz+2])[0]  # For current cell volume.
+        hz = np.diff(new_edges_z[iz:iz+2])[0]  # For current cell volume.
         for iy in range(ncy):
[…]
