+++ tests/test_utils.py
@@ -457,6 +457,84 @@ def test_get_receiver():
     assert_allclose(out7, 0.+0j)
 
 
+def test_grid2grid_volume():
+    # == X == Simple 1D model
+    grid_in = utils.TensorMesh(
+            [np.ones(5)*10, np.array([1, ]), np.array([1, ])],
+            x0=np.array([0, 0, 0]))
+    grid_out = utils.TensorMesh(
+            [np.array([10, 25, 10, 5, 2]), np.array([1, ]), np.array([1, ])],
+            x0=np.array([-5, 0, 0]))
+    values_in = np.array([1., 5., 3, 7, 2])[:, None, None]
+    values_out = utils.grid2grid(grid_in, values_in, grid_out)
+
+    # Result 2nd cell: (5*1+10*5+10*3)/25=3.4
+    assert_allclose(values_out[:, 0, 0], np.array([1, 3.4, 7, 2, 2]))
+
+    # == Y ==  Reverse it
+    grid_out = utils.TensorMesh(
+            [np.array([1, ]), np.ones(5)*10, np.array([1, ])],
+            x0=np.array([0, 0, 0]))
+    grid_in = utils.TensorMesh(
+            [np.array([1, ]), np.array([10, 25, 10, 5, 2]), np.array([1, ])],
+            x0=np.array([0, -5, 0]))
+    values_in = np.array([1, 3.4, 7, 2, 2])[None, :, None]
+    values_out = utils.grid2grid(grid_in, values_in, grid_out)
+
+    # Result 1st cell: (5*1+5*3.4)/10=2.2
+    assert_allclose(values_out[0, :, 0], np.array([2.2, 3.4, 3.4, 7, 2]))
+
+    # == Z == Another 1D test
+    grid_in = utils.TensorMesh(
+            [np.array([1, ]), np.array([1, ]), np.ones(9)*10],
+            x0=np.array([0, 0, 0]))
+    grid_out = utils.TensorMesh(
+            [np.array([1, ]), np.array([1, ]), np.array([20, 41, 9, 30])],
+            x0=np.array([0, 0, 0]))
+    values_in = np.arange(1., 10)[None, None, :]
+    values_out = utils.grid2grid(grid_in, values_in, grid_out)
+
+    assert_allclose(values_out[0, 0, :], np.array([1.5, 187/41, 7, 260/30]))
+
+    # == 3D ==
+    grid_in = utils.TensorMesh(
+            [np.array([1, 1, 1]), np.array([10, 10, 10, 10, 10]),
+             np.array([10, 2, 10])], x0=np.array([0, 0, 0]))
+    grid_out = utils.TensorMesh(
+            [np.array([1, 2, ]), np.array([10, 25, 10, 5, 2]),
+             np.array([4, 4, 4])], x0=np.array([0, -5, 6]))
+    create = np.array([[1, 2., 1]])
+    create = np.array([create, 2*create, create])
+    values_in = create*np.array([1., 5., 3, 7, 2])[None, :, None]
+
+    values_out = utils.grid2grid(grid_in, values_in, grid_out)
+
+    check = np.array([[1, 1.5, 1], [1.5, 2.25, 1.5]])[:, None, :]
+    check = check*np.array([1, 3.4, 7, 2, 2])[None, :, None]
+
+    assert_allclose(values_out, check)
+
+    # == If the extent is the same, volume*values must remain constant. ==
+    grid_in = utils.TensorMesh(
+            [np.array([1, 1, 1]), np.array([10, 10, 10, 10, 10]),
+             np.array([10, 2, 10])], x0=np.array([0, 0, 0]))
+    grid_out = utils.TensorMesh(
[…]
