### 2019-07-13 13a0827
Better commenting volume average
