+++ emg3d/njitted.py
@@ -1999,11 +1999,12 @@ def restrict_weights(vectorN, vectorCC, h, cvectorN, cvectorCC, ch):
 
 # Volume averaging
 @nb.njit(**_numba_setting)
-def volume_average(edges_x, edges_y, edges_z, values, new_edges_x, new_edges_y,
-                   new_edges_z, new_values):
+def volume_average(edges_x, edges_y, edges_z, values,
+                   new_edges_x, new_edges_y, new_edges_z, new_values):
     """Interpolation using the volume averaging technique.
 
     The result is added to new_values.
+
 
     Parameters
     ----------
@@ -2035,20 +2036,22 @@ def volume_average(edges_x, edges_y, edges_z, values, new_edges_x, new_edges_y,
     ncy = len(new_edges_y)-1
     ncz = len(new_edges_z)-1
 
-    # Working arrays for edges; ensure they are big enough.
-    x_edges = np.zeros(max(ncx, len(edges_x))+3)
-    y_edges = np.zeros(max(ncy, len(edges_y))+3)
-    z_edges = np.zeros(max(ncz, len(edges_z))+3)
+    # Working arrays for edges.
+    x_edges = np.empty(len(edges_x)+2)
+    y_edges = np.empty(len(edges_y)+2)
+    z_edges = np.empty(len(edges_z)+2)
 
     # Loop over new_grid cells.
     for iz in range(ncz):
-        hz = np.diff(new_edges_z[iz:iz+2])[0]  # For current cell volume.
-        for iy in range(ncy):
-            hyz = hz*np.diff(new_edges_y[iy:iy+2])[0]  # "
-            for ix in range(ncx):
-                hxyz = hyz*np.diff(new_edges_x[ix:ix+2])[0]  # "
+        hz = np.diff(new_edges_z[iz:iz+2])[0]  # To calc. current cell volume.
 
-                # Get start cell and number of cells of grid.
+        for iy in range(ncy):
+            hyz = hz*np.diff(new_edges_y[iy:iy+2])[0]  # " "
+
+            for ix in range(ncx):
+                hxyz = hyz*np.diff(new_edges_x[ix:ix+2])[0]  # " "
+
+                # Get start edge and number of cells of original grid involved.
                 s_cx = ix_r[ix]
                 n_cx = ix_l[ix+1] - s_cx
 
@@ -2058,7 +2061,7 @@ def volume_average(edges_x, edges_y, edges_z, values, new_edges_x, new_edges_y,
                 s_cz = iz_r[iz]
                 n_cz = iz_l[iz+1] - s_cz
 
-                # Get the edge locations.
+                # Get the involved original grid edges for this cell.
                 x_edges[0] = new_edges_x[ix]
                 for i in range(n_cx):
                     x_edges[i+1] = edges_x[s_cx+i+1]
@@ -2074,17 +2077,23 @@ def volume_average(edges_x, edges_y, edges_z, values, new_edges_x, new_edges_y,
                     z_edges[k+1] = edges_z[s_cz+k+1]
                 z_edges[n_cz+1] = new_edges_z[iz+1]
 
-                # Calculate the cell value (times volume).
+                # Loop over each (partial) cell of the original grid which
+                # contributes to the current cell of the new grid and add its
+                # (partial) value.
                 for k in range(n_cz+1):
                     dz = np.diff(z_edges[k:k+2])[0]
+                    k += s_cz
+
                     for j in range(n_cy+1):
                         dyz = dz*np.diff(y_edges[j:j+2])[0]
+                        j += s_cy
+
                         for i in range(n_cx+1):
                             dxyz = dyz*np.diff(x_edges[i:i+2])[0]
+                            i += s_cx
 
[…]
