### 2019-07-15 f52de23
Make conjugate a parameter
