### 2019-07-15 21ffd84
Improve stretched grid if x1
