### 2019-07-15 582f875
Work on get_receiver; docs/tests missing
