### 2019-07-15 1df52fe
Improve interpolation; tests and docs
