+++ emg3d/utils.py
@@ -1156,13 +1156,11 @@ def get_h_field(grid, model, field):
     return hfield
 
 
-def get_receiver(grid, fieldf, rec_loc, method='cubic'):
-    """Return field at receiver locations.
+def get_receiver(grid, values, coordinates, method='cubic', extrapolate=False):
+    """Return values corresponding to grid at coordinates.
 
     Works for electric fields as well as magnetic fields obtained with
-    :func:`get_h_field`.
-
-    If ``rec_loc`` is outside of ``grid``, the returned field is zero.
+    :func:`get_h_field`, and for model parameters.
 
 
     Parameters
@@ -1170,111 +1168,75 @@ def get_receiver(grid, fieldf, rec_loc, method='cubic'):
     grid : TensorMesh
         Model grid; a ``TensorMesh``-instance.
 
-    fieldf : ndarray
-        Field in a direction, e.g., for the x-direction ``Field.fx``, with
-        shape ``Field.vnEx``.
+    values : ndarray
+        Can be either a particular field, e.g., efield.fx, or a model
+        parameter.
 
-    rec_loc : tuple (rec_x, rec_y, rec_z)
-        ``rec_x``, ``rec_y``, and ``rec_z`` positions.
+    coordinates : tuple (x, y, z)
+        Coordinates (x, y, z) where to interpolate ``values``; e.g. receiver
+        locations.
 
     method : str, optional
         The method of interpolation to perform, 'linear' or 'cubic'.
         Default is 'cubic' (forced to 'linear' if there are less than 3 points
         in any direction).
 
+    extrapolate : bool
+        If True, points on ``new_grid`` which are outside of ``grid`` are
+        filled by the nearest value (if ``method='cubic'``) or by extrapolation
+        (if ``method='linear'``). If False, points outside are set to zero.
+
+        Default is False.
+
 
     Returns
     -------
-    values : ndarray
-        ``fieldf`` at positions ``rec_loc``.
+    new_values : ndarray
+        Values at ``coordinates``.
+
+
+    See Also
+    --------
+    grid2grid : Interpolation of model parameters or fields to a new grid.
 
     """
-
-    # TODO          TODO          TODO
-    #
-    # Generalized it so
-    # - rec_loc can also be a grid.
-    # - fieldf can also be a Field-instance, in which case fx, fy, and fz are
-    #   calculated recursively and a Field-instance is returned.
-    #
-    # => Documentation and Tests are missing.
-    #
-    # => Combine grid2grid and get_receiver?
-    #
-    # TODO          TODO          TODO
-
-    if hasattr(fieldf, 'fx'):
-        fx = get_receiver(grid, np.asarray(fieldf.fx), rec_loc, method)
-        fy = get_receiver(grid, np.asarray(fieldf.fy), rec_loc, method)
-        fz = get_receiver(grid, np.asarray(fieldf.fz), rec_loc, method)
-
-        return Field(fx, fy, fz)
-
     # Ensure input field is a certain field, not a Field instance.
-    if fieldf.ndim == 1:
-        print("* ERROR   :: Field must be x-, y-, or z-directed with ndim=3.")
-        print(f"             Shape of provided field: {fieldf.shape}.")
-        raise ValueError("Field error")
+    if values.ndim == 1:
+        print("* ERROR   :: Values must be a x-, y-, or z-directed field or")
+        print("             model parameters with ndim=3.")
+        print(f"             Shape of provided field: {values.shape}.")
+        raise ValueError("Values error")
 
-    # Check if rec_loc is a grid or a tuple/list.
-    if hasattr(rec_loc, 'hx'):
-        calc_new_points = True
-        new_points = tuple()
-    else:
-        calc_new_points = False
[…]
