### 2019-07-15 #1463 @akelbert
Hi @prisae I apologize for dragging this review out like I did. I have now been able to run several, although not all, of the Jupyter notebook examples successfully. The greatest challenge so far has been in ensuring that all of the dependencies are installed, and there are many.

Let me provide context. To run even the simplest 0a_Minimum_working_example, discretize package was needed. With the complicated security/SSL certificates at USGS, this involved cloning each and every one of the dependencies and compiling locally. However, the simple "python setup.py install" did not work for discretize, as advertised on https://pypi.org/project/discretize/.

Instead, the following worked
`
python setup.py build_ext --inplace

then add this to ~/.bashrc:

export PYTHON_PATH=~/Developer/local/anaconda2019/lib/python3.7/site-packages/discretize-0.4.8-py3.7-macosx-10.7-x86_64.egg/discretize/:$PYTHON_PATH`

Once the background infrastructure was installed, I was pleasantly surprised with the modularity and intuitiveness of emg3d interfaces. However, I found it difficult to deviate from the example scripts, and new dependencies kept crawling upon me with every new example. For the same reason, I have decided to call this good in terms of my time investment and have not proceeded to set up realistic nested grid examples to validate against other modeling tools. I trust the authors to have done that.

To sum up, I am sure that this is a great tool for a specialist already embedded in the python infrastructure. For someone who is a working EM researcher but who isn't part of the ecosystem just yet, the price to pay to get started is a little high at present. I recommend a start up guide that lists all the dependencies upfront and also provides some high level documentation on the specific utilities of the modules. Other than that, I am very impressed with this development and recommend it for publication. Sorry that I couldn't be more helpful on the technical aspects of the code.
