### 2019-07-16 #1463 @prisae
Thanks @akelbert for your review! It is interesting and insightful to have feedback from someone who is not familiar with the Python ecosystem.

`emg3d` itself has very few hard dependencies, only `numpy`, `scipy`, and `numba`. The former two are the core of the scientific Python stack, leaving the last one as only 'hurdle'. (As soon as you want to plot something you naturally also need `matplotlib` or another plotting library.) However, `emg3d` in itself is "only" a solver. It does not include plotting functions, model creation tools etc., there are other packages which do that. This is one of the ideas in the emg3d-examples-notebooks. They show how you can use `discretize` to create advanced meshes and plot the result, how you can use `GemPy` to generate a geological model and then calculate the CSEM response using `emg3d`, how to use it with `PyVista` and `Panel` to have fancy interactive 3D plots within the notebook, compare it to other modellers such us `empymod` or `MARE2DEM`, or how to use it with the `SimPEG` framework etc. All these additional packages obviously come with their own dependency, sometimes a lot as you had to learn the hard way. Usually all of them can be installed with a simple `pip install` or `conda install` though, unless there are firewall issues...

So I agree with you that the biggest hurdle is probably the ecosystem, as I think the issues you raised are not related to `emg3d`, but to the usage of Python in general.

> I recommend a start up guide that lists all the dependencies upfront [...]

`emg3d` has as hard dependency only `numpy`, `scipy`, and `numba`, as listed on the website and in the documentation. Most notebooks have a **Requires** section before the first cell with the imports. Also, all notebooks have all imports in the first cell, so from looking at the first cell you know the requirements for that particular notebook. Do you think I should add more to the docs than that?

> [...] and also provides some high level documentation on the specific utilities of the modules.

There are these two pages:
- https://emg3d.readthedocs.io/en/stable/usage.html
- https://emg3d.readthedocs.io/en/stable/solver.html

If you think that there should be more high level documentation I'd be happy to add more. Could you please elaborate a bit in this case?
