### 2019-07-16 #1463 @prisae
I will add to the `0a_Minimum_working_example` some more text with the following example which indeed just requires emg3d (with its hard dependencies `numpy`, `scipy`, and `numba`):

```py
import emg3d
import numpy as np

# Create a simple grid, 8 cells of length 1 in each direction, starting at the origin.
grid = emg3d.utils.TensorMesh([np.ones(8), np.ones(8), np.ones(8)], x0=np.array([0, 0, 0]))

# Define model frequency (Hz).
freq = 10.0

# The model is a fullspace with tri-axial anisotropy.
model = emg3d.utils.Model(grid, res_x=1.5, res_y=1.8, res_z=3.3, freq=freq)

# The source is a x-directed, horizontal dipole at (4, 4, 4).
sfield = emg3d.utils.get_source_field(grid, src=[4, 4, 4, 0, 0], freq=freq)

# Calculate the electric signal.
efield = emg3d.solver.solver(grid, model, sfield, verb=3)

# Get the corresponding magnetic signal.
hfield = emg3d.utils.get_h_field(grid, model, efield)
```
Example output:
```bash

:: emg3d START :: 07:02:15 ::

   MG-cycle       : 'F'                 sslsolver : False
   semicoarsening : False [0]           tol       : 1e-06
   linerelaxation : False [0]           maxit     : 50
   nu_{i,1,c,2}   : 0, 2, 1, 2          verb      : 3
   Original grid  :   8 x   8 x   8     => 512 cells
   Coarsest grid  :   2 x   2 x   2     => 8 cells
   Coarsest level :   2 ;   2 ;   2

   [hh:mm:ss]  rel. error                  [abs. error, last/prev]   l s

       h_
      2h_ \    /
      4h_  \/\/

   [07:02:15]   2.284e-02  after   1 F-cycles   [1.275e-06, 0.023]   0 0
   [07:02:15]   1.565e-03  after   2 F-cycles   [8.739e-08, 0.069]   0 0
   [07:02:15]   1.295e-04  after   3 F-cycles   [7.232e-09, 0.083]   0 0
   [07:02:15]   1.197e-05  after   4 F-cycles   [6.685e-10, 0.092]   0 0
   [07:02:15]   1.233e-06  after   5 F-cycles   [6.886e-11, 0.103]   0 0
   [07:02:15]   1.415e-07  after   6 F-cycles   [7.899e-12, 0.115]   0 0

   > CONVERGED
   > MG cycles        : 6
   > Final rel. error : 1.415e-07

:: emg3d END   :: 07:02:15 :: runtime = 0:00:00
```

I hope this example will make it even easier to get started.
