### 2019-07-16 3d89284
Change default in grid2grid
grid2grid is now more general, and can also be used for fields. However,
`method='volume'` is not implemented for fields (and does not make
sense), whereas 'cubic' is often not a good idea for model parameters
(or should be done on a log-scale). So the default now is 'linear'.
While often not the best, it is the most robust and the fastest.
