### 2019-07-16 57a0ee1
Interpolation and other improvements
- Make conjugate a parameter in `solver.solver`.
- Improve stretched grid if `x1` in `utils.get_stretched_h`.
- Improve `utils.get_receiver` and `utils.grid2grid`.
- Change default in `utils.grid2grid`: grid2grid is now more general, and can also be used for fields. However, `method='volume'` is not implemented for fields (and does not make sense), whereas 'cubic' is often not a good idea for model parameters (or should be done on a log-scale). So the default now is 'linear'. While often not the best, it is the most robust and the fastest.
