### 2019-07-16 3eb10a0
Add exit status to info_dict
