+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
   volume-average technique.
 - New parameter ``conjugate`` in ``solver.solver`` to permit both Fourier
   transform conventions.
+- Added ``exit`` and ``exit_message`` to ``info_dict``.
 
 
 *v0.7.0* : H-field
