### 2019-07-16 bbc89a4
Add exit status to info_dict
