### 2019-07-16 #1463 @akelbert
@prisae Your level of documentation goes as far as open source software documentation could reasonably be expected to go without putting an undue burden on the contributors of the software. I would only add a page that details which packages are "Recommended" (rather than "Required") to obtain the full functionality of emg3d, and the "Related" packages that you would want to also consider, e.g., for validation such as empymod. I would write a short paragraph about each package and how it's related to the functionality of yours. Make that a high-level header that the users could refer to upfront to understand all that's involved in using your package. I don't have any additional comments on this review, so please consider it completed.
