### 2019-07-16 a9080bb
Correct typos JOSS paper
