+++ docs/paper.bib
@@ -1,7 +1,7 @@
 @BOOK{Briggs:2000,
   AUTHOR    = {W. Briggs and V. Henson and S. McCormick},
   TITLE     = {{A Multigrid Tutorial, Second Edition}},
-  PUBLISHER = {Society for Industrial and Applied Mathematics},
+  PUBLISHER = {{Society for Industrial and Applied Mathematics}},
   YEAR      = {2000},
   DOI       = {10.1137/1.9780898719505}
 }
