### 2019-07-16 #1463 @prisae
I revised the manuscript and corrected some typos @jedbrown ; from my point of view it is good to go.

One question though: I realized I did not introduce the variables of the provided equation. Should I add this, or is it out of scope for JOSS?

I will add a section to the docs now as suggested by @akelbert
