### 2019-07-16 #1463 @prisae
I added a short section **Related Ecosystem** in the **Getting Started**-guide just below the **Example** section, https://emg3d.readthedocs.io/en/latest/usage.html#related-ecosystem

It is not very extensive at the moment, but it might expand with time.

Also, I don't want to dig into dependencies of these related packages. One would have to consult these packages to see their dependencies.
