### 2019-07-16 #1463 @jedbrown
Yes, I think the paper should be self-contained in the sense of defining the variables used in the equation. Some editorial suggestions:

* You have a stray `)` after the Werthmuller reference.
* I think the semicolon list should use commas.
* FEniCS and Pardiso may warrant a citation.
* Grammar: "several millions of cells" -> "several million cells"
* `Mulder [@Mulder:2006]` should be `@Mulder:2006` (see "in-text citation" at https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)
* `[@Mulder:2006]; [@Mulder:2007]` should be `[@Mulder:2006;@Mulder:2007]`
* No hyphen in `SciPy-stack` and it warrants a citation (https://scipy.org/citing.html)
* Also Numba (https://numba.pydata.org/numba-doc/latest/user/faq.html#how-do-i-reference-cite-acknowledge-numba-in-other-work)

Finally, what do you think about showing something to characterize performance relative to direct solvers? A potential user not familiar with solver complexity and multigrid methods won't see a compelling reason to care about emg3d just by reading this paper. A figure showing improved scaling or some more tangible statement than the "require substantial memory" sentence would help such a reader, IMO.
