+++ docs/paper.md
@@ -22,7 +22,7 @@ affiliations:
    index: 1
  - name: Shell Global Solutions International BV, Grasweg 31, 1031 HW Amsterdam, NL
    index: 2
-date: 03 May 2019
+date: 17 July 2019
 bibliography: paper.bib
 ---
 
@@ -42,38 +42,38 @@ $$ \eta \mathbf{E} - \nabla \times \mu_r^{-1} \nabla \times \mathbf{E} =
 where $\eta = \mathrm{i}\omega \mu_0(\sigma - \mathrm{i}\omega\varepsilon)$.
 
 Various open-source codes exist to model CSEM responses for a layered Earth,
-e.g., *DIPOLE1D* [@Key:2009] and *empymod* [@Werthmuller:2017], and for a
-two-dimensional Earth, e.g., *MARE2DEM* [@Key:2011]. Open-source modellers
-for a three-dimensional (3D) Earth only recently became available, notably
-*SimPEG* [@Cockett:2015]; *PETGEM* [@Castillo:2018]; and *custEM*
-[@Rochlitz:2019]. SimPEG is a framework that not only includes CSEM but also
-other geophysical methods and can model them on various types of regular grids.
-It currently uses mostly the direct solver *PARDISO*. PETGEM and custEM use
-finite elements with the *FEniCS* solver. All three codes require substantial
-memory and are not easily run on a laptop for models with several millions of
-cells.
+e.g., *DIPOLE1D* [@DIPOLE1D] and *empymod* [@empymod], and for a
+two-dimensional Earth, e.g., *MARE2DEM* [@MARE2DEM]. Open-source modellers for
+a three-dimensional (3D) Earth only recently became available, notably *SimPEG*
+[@SimPEG], *PETGEM* [@PETGEM], and *custEM* [@custEM]. SimPEG is a framework
+that not only includes CSEM but also other geophysical methods and can model
+them on various types of regular grids. It currently uses mostly the direct
+solver *PARDISO* [@PARDISO]. PETGEM and custEM use finite elements with the
+*FEniCS* [@FEniCS] solver. All three codes require substantial memory and are
+not easily run on a laptop for models with several million cells.
 
-Mulder [@Mulder:2006] has shown that the multigrid method [@Briggs:2000], which
-reduces memory requirement substantially, works fine for diffusive CSEM
-problems. This was later also confirmed by others [@Jaysaval:2016]. However, the
-multigrid CSEM codes of these publications are proprietary. The code *emg3d* is
-a multigrid solver for 3D CSEM diffusion with tri-axial electrical anisotropy
-using a staggered grid [@Mulder:2006]; [@Mulder:2007].
-It can act as a solver on its own, or be used as a preconditioner for various
-Krylov subspace methods included in the scientific Python library (SciPy).
-Multigrid solvers can struggle to converge with strong grid-stretching or
-strong anisotropy. We implemented *semicoarsening* (coarsening the grid only in
-some coordinate directions) and *line relaxation* (solving for the values of a
-whole gridline simultaneously in some directions) inside the multigrid
-technique to deal with these issues [@Jonsthovel:2006].
+@Mulder:2006 has shown that the multigrid method [@Briggs:2000], which reduces
+memory requirement substantially, works fine for diffusive CSEM problems. This
+was later also confirmed by others [@Jaysaval:2016]. However, the multigrid
+CSEM codes of these publications are proprietary. The code *emg3d* is a
+multigrid solver for 3D CSEM diffusion with tri-axial electrical anisotropy
+using a staggered grid [@Mulder:2006;@Mulder:2007]. It can act as a solver on
+its own, or be used as a preconditioner for various Krylov subspace methods
+included in the scientific Python library SciPy. Multigrid solvers can struggle
[…]
