### 2019-07-17 3ef1e53
Add a figure
