+++ docs/paper.bib
@@ -6,7 +6,7 @@
   DOI       = {10.1137/1.9780898719505}
 }
 
-@Misc{Scipy,
+@Misc{SciPy,
   AUTHOR    = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   TITLE     = {{SciPy}: Open source scientific tools for {Python}},
   YEAR      = {2001--},
@@ -73,7 +73,7 @@
   DOI       = {10.1190/1.3058434}
   }
 
-@ARTICLE{Numpy,
+@ARTICLE{NumPy,
   AUTHOR = {Stéfan van der Walt and S[] Chris Colbert and Gaël Varoquaux},
   TITLE = {The {NumPy} Array: A Structure for Efficient Numerical Computation},
   JOURNAL = {Computing in Science \& Engineering},
