### 2019-07-17 1ed33fe
Integrate figure better
