+++ docs/paper.bib
@@ -9,7 +9,7 @@
 @Misc{SciPy,
   AUTHOR    = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   TITLE     = {{SciPy}: Open source scientific tools for {Python}},
-  YEAR      = {2001--},
+  YEAR      = {2001},
   URL       = {http://www.scipy.org},
   NOTE      = {[Online; accessed 2019-07-17]}
 }
