### 2019-07-17 #1463 @prisae
Thanks for your comments @jedbrown
- I was wondering about how to better include the references, but couldn't find any info on the JOSS instructions; thanks for the link!
- Good point about references for SciPy/NumPy/Numba/FEniCS/PARDISO.

In order to talk about the runtime and memory of other solvers I would have to make more tests with other solvers. So I prefer to show a figure of the runtime and memory usage of emg3d, because that I tested extensively; I added a figure in this respect and adjusted the text accordingly.

I keep it in the branch `joss` (`@whedon generate pdf from branch joss`) for now, until you are happy with it.
