### 2019-07-17 98f147d
Add CPU info
