+++ docs/paper.md
@@ -57,7 +57,7 @@ solver *PARDISO* [@PARDISO]. PETGEM and custEM use finite elements with the
 not easily run on a laptop for models with several million cells.
 
 @Mulder:2006 has shown that the multigrid method [@Briggs:2000], with its
-optimal scaling for both runtime and memory consumption, as shown in Figure 1,
+optimal scaling for both runtime and memory consumption as shown in Figure 1,
 works fine for diffusive CSEM problems. This was later also confirmed by others
 [@Jaysaval:2016]. However, the multigrid CSEM codes of these publications are
 proprietary. The code *emg3d* is a multigrid solver for 3D CSEM diffusion with
@@ -70,7 +70,7 @@ directions) and *line relaxation* (solving for the values of a whole gridline
 simultaneously in some directions) inside the multigrid technique to deal with
 these issues [@Jonsthovel:2006].
 
-![Example showing the optimal scaling of emg3d for both runtime and memory consumption. Here for a homogeneous fullspace of 1 S/m and a frequency of 1 Hz.](paper.png)
+![Example showing the optimal scaling of emg3d for both runtime and memory consumption. Here for a homogeneous fullspace of 1 S/m and a frequency of 1 Hz (single thread on an Intel(R) Xeon(R) CPU @ 2.50GHz).](paper.png)
 
 The code is written completely in Python using the NumPy/SciPy stack
 [@NumPy;@SciPy], where the most time- and memory-consuming parts are sped up
