### 2019-07-17 #1463 @prisae
I added CPU info to figure.
