### 2019-07-17 #1463 @jedbrown
@prisae Thanks for your revisions. Do you think it would be appropriate to mention Hypre and ML/MueLu's Maxwell solvers? They use different MG formulations, but are open source, (distributed memory) parallel, and have fewer grid restrictions (good or bad depending on an application's needs). A sentence on why one would choose emg3d (e.g., improved robustness, Python, easier install?) would be helpful, IMO. Once you do that (or decide not to), please tag (presumably `v0.7.1`; [annotated tags](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred) and archive on Zenodo or similar, then report the DOI back here.
