### 2019-07-17 #1463 @prisae
Thanks @jedbrown, very interesting, I never heard of either of them. I will have a quick look and add them with a reference. Without going into details I will probably simply mention that there exist other MG packages which solve Maxwell's equations in other programming languages.
