### 2019-07-17 #1463 @prisae
If it is OK with you @jedbrown  I will not mention Hypre nor ML/MueLu. All the packages I mention (DIPOLE1D, empymod, MARE2DEM, PETGEM, SimPEG, custEM) are within the same realm, modelling geophysical controlled-source electromagnetic data (in the wider sense, not just CSEM, but also AEM, MT, etc.). However, I do not see how Hypre or ML/MueLu fit in, and from glancing at it and its examples it is not readily obvious to me how I could use them to model it, although it is most likely possible. But anything I would say about them would just be a guess from my side, and I don't feel comfortable with that.

The reason that I mention PARDISO and FEniCS is because they are used by packages which do CSEM calculations (SimPEG, PETGEM, and custEM). So if there would be examples/packages utilizing Hypre or ML/MueLu for CSEM modelling then indeed I should mention it, but I have not found anything. Do  you know of any such project, @jedbrown?
