### 2019-07-17 94e7331
Improve figure; last typos
