+++ docs/paper.py
@@ -1,4 +1,5 @@
 # JOSS figure
+# Python file to recreate the figure in the JOSS article paper.md.
 import numpy as np
 import matplotlib.pyplot as plt
 
@@ -8,46 +9,49 @@ plt.rcParams['savefig.dpi'] = 300
 plt.rcParams['font.size'] = 8
 
 # Data is from the notebooks `4a_RAM-requirements.ipynb` and `4b_Runtime.ipynb`
-# in the repo https://github.com/empymod/emg3d-examples.
+# in the repo https://github.com/empymod/emg3d-examples at commit 1a1a658c23
+# (2019-06-05). It was run at the TU Delft server Texel on one thread. From
+# `cat /proc/cpuinfo`: Intel(R) Xeon(R) CPU E5-2680 v3 @ 2.50GHz
 nsizes_mem = np.array([64, 96, 128, 192, 256, 384, 512, 768, 1024])
 mem = np.array([212, 379, 678, 1821, 4035, 13194, 30928, 103691, 245368])
 nsizes_cpu = np.array([32, 48, 64, 96, 128, 192, 256, 384])
 cpu = np.array([1.3, 4.5, 10.6, 36.3, 90.6, 311.9, 758.7, 2727.8])
 
+
+def focus_on_data(ax):
+    """Reduce and hide figure elements to focus on data."""
+    ax.grid(axis='y', color='grey', linestyle='-', linewidth=1, alpha=0.3)
+    ax.spines['top'].set_visible(False)
+    ax.spines['right'].set_visible(False)
+    ax.spines['bottom'].set_visible(False)
+    ax.spines['left'].set_visible(False)
+
+
 # Figure
-plt.figure(figsize=(8, 3))
+fig = plt.figure(figsize=(6, 2.5))
 
+# Runtime subplot
 ax1 = plt.subplot(121)
-plt.title('Time')
-plt.loglog(nsizes_cpu**3/1e6, cpu, '.-', label='runtime')
-plt.xlabel('Number of cells (in millions)')
-plt.ylabel('Runtime (s)')
+plt.title('(a) Time', fontsize=10)
+plt.loglog(nsizes_cpu**3/1e6, cpu, '.-')
+plt.ylabel('Runtime (s)', fontsize=8)
 plt.xticks([1e-1, 1e0, 1e1, 1e2], ('0.1', '1', '10', '100'))
-plt.yticks([1e0, 1e1, 1e2, 1e3, 1e4], ('1', '10', '100', '1000', '10000'))
+plt.yticks([1e0, 1e1, 1e2, 1e3], ('1', '10', '100', '1000'))
+focus_on_data(ax1)
 
-# Show horizontal gridlines, switch off frame
-ax1.grid(axis='y', color='grey', linestyle='-', linewidth=0.5, alpha=0.3)
-ax1.spines['top'].set_visible(False)
-ax1.spines['right'].set_visible(False)
-ax1.spines['bottom'].set_visible(False)
-ax1.spines['left'].set_visible(False)
-
+# Memory subplot
 ax2 = plt.subplot(122)
 ax2.yaxis.tick_right()
 ax2.yaxis.set_label_position("right")
-plt.title('Memory')
-plt.loglog(nsizes_mem**3/1e6, mem/1e3, '.-', zorder=10, label='MG full RAM')
-plt.xlabel('Number of cells (in millions)')
-plt.ylabel('RAM (GB)')
+plt.title('(b) Memory', fontsize=10)
+plt.loglog(nsizes_mem**3/1e6, mem/1e3, '.-')
+plt.ylabel('RAM (GB)', fontsize=8)
[…]
