### 2019-07-17 02e152d
Joss article
Add figure and more references; fix typos.
