+++ docs/paper.bib
@@ -6,6 +6,29 @@
   DOI       = {10.1137/1.9780898719505}
 }
 
+@Misc{SciPy,
+  AUTHOR    = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  TITLE     = {{SciPy}: Open source scientific tools for {Python}},
+  YEAR      = {2001},
+  URL       = {http://www.scipy.org},
+  NOTE      = {[Online; accessed 2019-07-17]}
+}
+
+@ARTICLE{PARDISO,
+  AUTHOR = {Olaf Schenk and Klaus Gärtner},
+  TITLE = {Solving unsymmetric sparse systems of linear equations with
+      PARDISO},
+  JOURNAL = {Future Generation Computer Systems},
+  VOLUME = {20},
+  NUMBER = {3},
+  PAGES = {475-487},
+  YEAR = {2004},
+  NOTE = {Selected numerical algorithms},
+  ISSN = {0167-739X},
+  DOI = {10.1016/j.future.2003.07.011},
+  URL = {http://www.sciencedirect.com/science/article/pii/S0167739X03001882},
+}
+
 @ARTICLE{Mulder:2006,
   AUTHOR    = {W. A. Mulder},
   TITLE     = {A multigrid solver for {3D} electromagnetic diffusion},
@@ -37,7 +60,7 @@
   DOI       = {10.3997/2214-4609.201401567}
 }
 
-@ARTICLE{Key:2009,
+@ARTICLE{DIPOLE1D,
   AUTHOR    = {K. Key},
   TITLE     = {{1D} inversion of multicomponent, multifrequency marine {CSEM}
                data: {M}ethodology and synthetic studies for resolving thin
@@ -50,10 +73,22 @@
   DOI       = {10.1190/1.3058434}
   }
 
-@ARTICLE{Key:2011,
+@ARTICLE{NumPy,
+  AUTHOR = {Stéfan van der Walt and S[] Chris Colbert and Gaël Varoquaux},
+  TITLE = {The {NumPy} Array: A Structure for Efficient Numerical Computation},
+  JOURNAL = {Computing in Science \& Engineering},
+  VOLUME = {13},
+  NUMBER = {2},
+  PAGES = {22-30},
+  YEAR = {2011},
+  DOI = {10.1109/MCSE.2011.37},
+  URL = {https://aip.scitation.org/doi/abs/10.1109/MCSE.2011.37},
+}
+
+@ARTICLE{MARE2DEM,
   AUTHOR    = {K. Key and J. Ovall},
-  TITLE     = {{A parallel goal-oriented adaptive finite element method for
-                2.5-D electromagnetic modelling}},
+  TITLE     = {A parallel goal-oriented adaptive finite element method for
+               {2.5-D} electromagnetic modelling},
   JOURNAL   = {Geophysical Journal International},
   YEAR      = {2011},
   VOLUME    = {186},
@@ -62,7 +97,38 @@
   DOI       = {10.1111/j.1365-246X.2011.05025.x}
 }
 
-@ARTICLE{Cockett:2015,
+@INPROCEEDINGS{Numba,
+  AUTHOR    = {Siu Kwan Lam and Antoine Pitrou and Stanley Seibert},
+  TITLE     = {{Numba: A LLVM-based Python JIT Compiler}},
+  BOOKTITLE = {Proceedings of the Second Workshop on the LLVM Compiler
+      Infrastructure in HPC},
+  SERIES = {LLVM '15},
+  YEAR = {2015},
+  ISBN = {978-1-4503-4005-2},
+  PAGES = {7:1--7:6},
+  ARTICLENO = {7},
+  NUMPAGES = {6},
[…]

[…]
