### 2019-07-17 #1463 @prisae
Here is the DOI: 10.5281/zenodo.3339421

`v0.7.1` will be on pypi and conda shortly.
