### 2019-07-17 #1463 @prisae
All done from my part. `v0.7.1` is released, it is on `pypi` and `conda`, and on Zenodo.

Thanks @jedbrown, @emersodb, @lukeolson, and @akelbert for your review and feedback!
