### 2019-07-17 #1463 @prisae
https://doi.org/10.5281/zenodo.3339421 now resolves, at least for me.

I adjusted the authors on Zenodo @jedbrown

![Selection_001](https://user-images.githubusercontent.com/8020943/61407380-e5e8f880-a8dd-11e9-9f99-67ea445eac65.png)
