### 2019-07-17 #1463 @jedbrown
@openjournals/joss-eics This paper is accepted; over to you.
