### 2019-07-17 #1463 @danielskatz
@prisae  - please change "A diffusive problem remains, which resulting system of equations is given in the frequency domain by" to "A diffusive problem remains, which has the resulting system of equations given in the frequency domain by"

"It currently uses mostly" should be "It currently mostly uses" or "It currently primarily uses"

"the multigrid CSEM codes of these publications" should be "the multigrid CSEM codes discussed in these publications"
