### 2019-07-18 #1463 @prisae
@danielskatz , does that imply that I'll have to make a new release and mint another Zenodo DOI with it? I am off on vacation right now, I'll do it when I am back on Monday.
