### 2019-07-18 #1463 @danielskatz
no, the paper source doesn't need to be in the repo or if it, it doesn't need to be up to date - we care about the software repo and the pdf of the paper in terms of archiving
