### 2019-07-22 1d9b303
Typos for JOSS
